% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{calJSI}
\alias{calJSI}
\title{calJSI}
\usage{
calJSI(maf, plot = TRUE, use.circle = TRUE)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function}

\item{plot}{logical, default TRUE}

\item{use.circle, }{default FALSE}
}
\value{
correlation matrix via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity coefficient is defined as the ratio of shared variants to all
(shared plus discordant) variants for two samples.
}
\examples{
calJSI(maf)
}
