% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{calMATH}
\alias{calMATH}
\title{mathScore}
\usage{
calMATH(VAF)
}
\arguments{
\item{maf}{a Maf object generated by readMaf.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{minvaf}{the minimum VAF for filtering variants. Default: 0.02 Option: on the scale of 0 to 1}

\item{maxvaf}{the maximum VAF for filtering variants. Default: 1. Option: on the scale of 0 to 1}

\item{plot}{TRUE(default)}
}
\value{
A list containing MATH scores and violin plot of all samples.
}
\description{
Calculate MATH score and VAF value could be filtered with minvaf and maxvaf.
}
\examples{
mathScore(maf, minvaf=0.02, maxvaf=1)
}
