% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNeiDist.R
\name{calNeiDist}
\alias{calNeiDist}
\title{calNeiDist}
\usage{
calNeiDist(
  maf,
  patient.id = NULL,
  withinTumor = FALSE,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.ccf = 0.02,
  plot = TRUE,
  use.circle = TRUE,
  title = NULL,
  number.cex = 8,
  number.col = "#C77960"
)
}
\arguments{
\item{maf}{An Maf object generated by readMaf function}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{withinTumor}{calculate fst within types in each patients,default is FALSE.}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, chrX, chrY. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSilent".}

\item{use.indel}{Logical. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.ccf}{specify the minimum CCF, default is 0.08}

\item{plot}{logical (Default:TRUE). Whether to show the plot, default TRUE}

\item{use.circle}{logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix}

\item{title}{the title of the plot, default is "Nei's distance"}
}
\value{
Nei's genetic distance matrix and heatmap of sample-pairs from the same patient
}
\description{
Nei's distance of CCF for each sample-pair
}
\examples{
calNeiDist(maf)
}
\references{
Lee JK, Wang J, Sa JK, et al. Spatiotemporal genomic architecture informs precision oncology in glioblastoma. Nat Genet. 2017;49(4):594-599.
}
