% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTMB.R
\name{calTumorBurden}
\alias{calTumorBurden}
\title{getTMB}
\usage{
calTumorBurden(vafColumnMATH)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{tsb}{Specify Tumor_Sample_Barcodes for single or all samples from maf of the patient. Default: NULL, which means that output all samples' TMB Score.}

\item{minvaf}{Default: 0. Option: on the scale of 0 to 0.02. Filter low frequency variants.}

\item{maxvaf}{Default: 0. Option: on the scale of 0 to 1. Filter extremely high frequency variants.}
}
\value{
a list containing MATH scores of all/selected samples.
}
\description{
Calculate tumor mutational burden(TMB) and VAF value could be filtered with minvaf and maxvaf.
}
