% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfAUC.R
\name{ccfAUC}
\alias{ccfAUC}
\title{ccfAUC}
\usage{
ccfAUC(maf, tsb = NULL)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{tsb}{specify Tumor_Sample_Barcodes for single or all samples from maf of the patient. Default: NULL, which means that output all samples' AUC Score.}
}
\value{
A list containing AUC scores' data frame and corresponding ccf/proporation plot.
}
\description{
The tumor heterogeneity was estimated as the area under the curve (AUC) of the cumulative density 
function from all cancer cell fractions per tumor
}
\examples{
ccfAUC(maf, tsb=NULL)
ccfAUC(maf, tsb=c("SRR3670035"))

}
