% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfAUC.R
\name{ccfAUC.sample}
\alias{ccfAUC.sample}
\title{ccfAUC}
\usage{
ccfAUC.sample(ccf)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.ccf}{the minimum value of CCF. Default: 0}

\item{plot.density}{whether to show the density plot. Default: TRUE}
}
\value{
A list containing AUC of CCF and a graph
}
\description{
The tumor heterogeneity was estimated as the area under the curve (AUC) of the cumulative density 
function from all cancer cell fractions per tumor
}
\examples{
ccfAUC(maf)
ccfAUC(maf, patient.id = "HCC4064")

}
\references{
Charoentong P, Finotello F, et al. Pan-cancer Immunogenomic Analyses Reveal Genotype-Immunophenotype Relationships and Predictors of Response to Checkpoint Blockade. 
Cell reports 2017, 18:248-262.
}
