% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfDensity.R
\name{ccfDensity}
\alias{ccfDensity}
\title{ccfDensity}
\usage{
ccfDensity(maf, show.density = TRUE)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{show.density}{logical. If TRUE, the default, perform density estimation.}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function is only available when the Maf object was generated along with CCF data.
}
\examples{
ccfDensity(maf, show.density = T)
}
