% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{ccfPair}
\alias{ccfPair}
\title{compareCCF}
\usage{
ccfPair(mafData, pairByType = FALSE, geneList = NULL, show.density = TRUE)
}
\arguments{
\item{pairByType}{pair by tumor types in each patients,default is FALSE.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{show.density}{logical. If TRUE, the default, perform density estimation.}

\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.ccf}{the minimum value of CCF. Default: 0}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
compareCCF(maf, show.density = TRUE)
}
