% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkMafInput}
\alias{checkMafInput}
\title{Check Maf or MafList input}
\usage{
checkMafInput(object, patient.id = NULL)
}
\arguments{
\item{object}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}
}
\value{
A list of Maf
}
\description{
Check Maf or MafList input
}
\examples{
checkMafInput(maf)
}
