% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkPhyloTreeInput}
\alias{checkPhyloTreeInput}
\title{Check phyloTree or phyloTreeList input}
\usage{
checkPhyloTreeInput(object, patient.id = NULL)
}
\arguments{
\item{object}{phyloTree or phyloTreeList object generated by \code{\link{getPhyloTree}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}
}
\value{
a list of phyloTree
}
\description{
Check phyloTree or phyloTreeList input
}
\examples{
checkPhyloTreeInput(phyloTree)

}
