% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.maf.R
\docType{class}
\name{classMaf-class}
\alias{classMaf-class}
\alias{classMaf}
\title{Output an new maf document with information of sample_info}
\arguments{
\item{patientID}{patient/sample name}

\item{dat.dir}{specify a data directory as the input of the function}

\item{use.ccf=Default}{FALSE. TRUE if ccf data is provided}

\item{plot.mafSummary}{Default TRUE. FALSE if the summary figure is 
unnecessary}

\item{ref.build}{Default hg19. specify a referential genome including 
BSgenome.Hsapiens.UCSC.hg19 and BSgenome.Hsapiens.UCSC.hg38}
}
\value{
a classMaf object/class includes information of sample_info and 
mut.id and summary figure of it
}
\description{
Add sample_info to the original maf file to get a new maf.The
 new maf file adds three pieces of information:lesion,patient and time
}
\examples{
\dontrun{
read.maf(patientID, dat.dir = "./data", use.ccf = FALSE, plot.mafSummary = TRUE) 
}
}
