% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{compareCCF}
\alias{compareCCF}
\title{compareCCF}
\usage{
compareCCF(maf, patient.id = NULL, pairByType = FALSE, min.ccf = 0)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{pairByType}{pair by tumor types in each patients,default is FALSE.}

\item{min.ccf}{the minimum value of CCF. Default: 0}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
compareCCF(maf)
}
