% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareJSI.R
\name{compareJSI}
\alias{compareJSI}
\title{compareJSI}
\usage{
compareJSI(
  maf,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  pairByTumor = FALSE,
  min.vaf = 0.02,
  plot = TRUE,
  use.circle = TRUE,
  title = NULL,
  number.cex = 8,
  number.col = "#C77960"
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, chrX, chrY. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.vaf}{Specify the minimum VAF_adj, default is 0.08}

\item{plot}{Logical (Default:TRUE).}

\item{use.circle}{Logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix}

\item{title}{Title of the plot, default is "Jaccard similarity"}
}
\value{
correlation matrix and heatmap via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity index (JSI) is applied to distinguish monoclonal versus polyclonal seeding in metastases
}
\examples{
compareJSI(maf)
}
\references{
Hu, Z., Li, Z., Ma, Z. et al. Multi-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. Nat Genet (2020).
}
