% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumberFilter.R
\name{copyNumberFilter}
\alias{copyNumberFilter}
\title{copyNumberFilter}
\usage{
copyNumberFilter(maf, seg)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{seg}{object generated by readSegment function.}
}
\description{
Ignoring variants in copy number altered regions
}
\examples{
copyNumberFiler(maf, seg)

}
