% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSignatures.R
\name{fitSignatures}
\alias{fitSignatures}
\title{fitSignatures}
\usage{
fitSignatures(
  tri_matrix = NULL,
  signaturesRef = "cosmic_v2",
  associated = NULL,
  min.mut.count = 15,
  signature.cutoff = 0.1,
  patient.id = NULL
)
}
\arguments{
\item{tri_matrix}{a matrix or a list of matrix generated by triMatrix function.}

\item{signaturesRef}{Signature reference,Users can upload their own reference. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{associated}{Associated Vector of associated signatures. If given, will narrow the signatures reference to only the ones listed.Default NULL.}

\item{min.mut.count}{The threshold for the variants in a branch. Default 15.}

\item{signature.cutoff}{Discard any signature contributions with a weight less than this amount.Default: 0.1.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}
}
\value{
A list of data frames, each one contains treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Find nonnegative linear combination of mutation signatures to reconstruct matrix and 
calculate cosine similarity based on somatic SNVs.
}
\examples{
fitSignatures(tri_matrix)
}
