% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFst.R, R/calculateFst.R
\name{fst.hudson.pair}
\alias{fst.hudson.pair}
\title{calFst}
\usage{
fst.hudson.pair(maf.pair)

fst.hudson.pair(maf.pair)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{use.adjVAF}{whether to use adjusted VAF values rather than original VAF values, default FALSE}

\item{min.vaf}{specify the minimum VAF, default is 0.08}

\item{max.vaf}{specify the minimum VAF, default is 1}
}
\value{
a list contains Fst value of MRS and Hudson estimator of each sample-pair, respectively.

a list contains Fst value of MRS and Hudson estimator of each sample-pair, respectively.
}
\description{
genetic divergence between regions of subclonal sSNVs using the Weir and Cockerham method

genetic divergence between regions of subclonal sSNVs using the Weir and Cockerham method
}
\examples{
calFst(maf)
calFst(maf)
}
\references{
Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. 
Nat Genet. 2017;49(7):1015–1024. doi:10.1038/ng.3891

Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and interpreting FST: the impact of rare variants. 
Genome Res. 2013;23(9):1514–1521. doi:10.1101/gr.154831.113

Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. 
Nat Genet. 2017;49(7):1015–1024. doi:10.1038/ng.3891

Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and interpreting FST: the impact of rare variants. 
Genome Res. 2013;23(9):1514–1521. doi:10.1101/gr.154831.113
}
