% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFst.R
\name{fst.hudson}
\alias{fst.hudson}
\title{calFst}
\usage{
fst.hudson(maf.pair, minVAF)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{min.VAF}{specify the minimum CCF value of subclonal mutations.}
}
\value{
a list contains Fst value of MRS and Hudson estimator of each sample-pair, respectively.
}
\description{
genetic divergence between regions of subclonal sSNVs using the Weir and Cockerham method
}
\examples{
calFst(maf)
}
\references{
Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. 
Nat Genet. 2017;49(7):1015–1024. doi:10.1038/ng.3891

Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and interpreting FST: the impact of rare variants. 
Genome Res. 2013;23(9):1514–1521. doi:10.1101/gr.154831.113
}
