% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutProfile_V1.R
\name{genHeatmapPlotMatrix}
\alias{genHeatmapPlotMatrix}
\title{plotMutProfile}
\usage{
genHeatmapPlotMatrix(maf_data, topGenesCount = NULL)
}
\arguments{
\item{topGenesCount}{the number of genes print, default is 10}

\item{maf}{the Maf object generated by readMaf}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{class}{classify SSNVs/Indels into Shared/P-shared/Private("SP")/, Clonal/Subclonl ("CS")
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal ("SPCS")
,default is "SP"}

\item{title}{the first half of the plot's title, default is "Mutational profile"}

\item{driverGenesFile}{driver gene list. Default NULL}
}
\description{
plotMutProfile
}
\examples{
plotMutProfile(maf, class = "SP", topGenesCount = 10)
}
