% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getBootstrapValue}
\alias{getBootstrapValue}
\alias{getBootstrapValue,phyloTree-method}
\title{getBootstrapValue}
\usage{
getBootstrapValue(object)

\S4method{getBootstrapValue}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Bootstrap value of phyloTree
}
\description{
getBootstrapValue
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile = maf.File, ccfFile = ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
getBootstrapValue(phyloTree)
}
