% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getBranchType}
\alias{getBranchType}
\alias{getBranchType,phyloTree-method}
\title{getBranchType}
\usage{
getBranchType(object)

\S4method{getBranchType}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Branch type of phyloTree
}
\description{
getBranchType
}
\examples{
getBranchType(phyloTree)
}
