% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCCFMatrix}
\alias{getCCFMatrix}
\alias{getCCFMatrix,phyloTree-method}
\title{getCCFMatrix}
\usage{
getCCFMatrix(object)

\S4method{getCCFMatrix}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
CCF matrix of phyloTree
}
\description{
getCCFMatrix
}
\examples{
getCCFMatrix(phyloTree)
}
