% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMafData}
\alias{getMafData}
\alias{getMafData,Maf-method}
\title{getMafData}
\usage{
getMafData(object)

\S4method{getMafData}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Maf data
}
\description{
getMafData
}
\examples{
getMafData(maf)
}
