% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMafPatient}
\alias{getMafPatient}
\alias{getMafPatient,Maf-method}
\title{getMafPatient}
\usage{
getMafPatient(object)

\S4method{getMafPatient}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Human reference genome versions of Maf
}
\description{
getMafPatient
}
\examples{
getMafPatient(maf)
}
