% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMafRef}
\alias{getMafRef}
\alias{getMafRef,Maf-method}
\title{getMafRef}
\usage{
getMafRef(object)

\S4method{getMafRef}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Human reference genome versions of Maf
}
\description{
getMafRef
}
\examples{
getMafRef(maf)
}
