% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMergedCCF.R
\name{getMergedCCF}
\alias{getMergedCCF}
\title{getMergedCCF}
\usage{
getMergedCCF(mafData)
}
\arguments{
\item{mafData}{the data information of the Maf object generated by readMaf function.}
}
\value{
a data.frame of merged CCF of sSNVs.
}
\description{
calculate merged CCF of each sSNV by integrating CCFs of MRS data.
}
\examples{
getMergedCCF(maf)
}
\references{
Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. 
Nat Genet. 2017;49(7):1015–1024. doi:10.1038/ng.3891
}
