% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMutBranches}
\alias{getMutBranches}
\alias{getMutBranches,phyloTree-method}
\title{getMutBranches}
\usage{
getMutBranches(object)

\S4method{getMutBranches}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
branches mutation of phyloTree
}
\description{
getMutBranches
}
\examples{
getMutBranches(phyloTree)
}
