% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJtree.R
\name{getNJtree}
\alias{getNJtree}
\title{getNJtree}
\usage{
getNJtree(maf, minVaf = 0.02, maxVaf = 1,
  ccf.mutation.id = c("Hugo_Symbol", "Chromosome", "Start_Position"),
  ccf.mutation.sep = ":")
}
\arguments{
\item{minVaf}{the minimum value of vaf}

\item{maxVaf}{the maximum value of vaf}

\item{ccf.mutation.id}{manually specify which columns could be joint by ccf.mutation.sep to get the same format of mutation id in ccfy}

\item{ccf.mutation.sep}{manually specify the separator character.Values on each line of the ccf.mutation.id are separated by this character. I}

\item{a}{Maf object generated by readMaf function}
}
\value{
an object of class NJtree
}
\description{
getNJtree
}
\examples{
maf.File <- system.file("extdata/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/ccf", "311252.loci.tsv", package = "Meskit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfoFile, refBuild="hg19")
njtree <- getNJtree(maf)
}
