% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getNonSyn_vc}
\alias{getNonSyn_vc}
\alias{getNonSyn_vc,Maf-method}
\title{getNonSyn_vc}
\usage{
getNonSyn_vc(object)

\S4method{getNonSyn_vc}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
A list of Variant classifications which are considered as non-silent.
}
\description{
getNonSyn_vc
}
\examples{
maf.File <- system.file("extdata/", "HCC_LDC.maf", package = "MesKit")
clin.File <- system.file("extdata/", "HCC_LDC.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
getNonSyn_vc(maf$HCC5647)
}
