% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPatientID}
\alias{getPatientID}
\alias{getPatientID,phyloTree-method}
\title{getPatientID}
\usage{
getPatientID(object)

\S4method{getPatientID}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
patientID of phyloTree
}
\description{
getPatientID
}
\examples{
getPatientID(phyloTree)
}
