% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPhyloTreePatient}
\alias{getPhyloTreePatient}
\alias{getPhyloTreePatient,phyloTree-method}
\title{getPhyloTreePatient}
\usage{
getPhyloTreePatient(object)

\S4method{getPhyloTreePatient}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
patientID of phyloTree
}
\description{
getPhyloTreePatient
}
\examples{
getPhyloTreePatient(phyloTree)
}
