% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPhyloTreeRef}
\alias{getPhyloTreeRef}
\alias{getPhyloTreeRef,phyloTree-method}
\title{getPhyloTreeRef}
\usage{
getPhyloTreeRef(object)

\S4method{getPhyloTreeRef}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
reference genome versions of phyloTree
}
\description{
getPhyloTreeRef
}
\examples{
getPhyloTreeRef(phyloTree)
}
