% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPhyloTreeTsbLabel}
\alias{getPhyloTreeTsbLabel}
\title{getPhyloTreeRef}
\usage{
getPhyloTreeTsbLabel(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
relationship between Tumor_Sample_Barcode and Tumor_Sample_Label
}
\description{
getPhyloTreeRef
}
\examples{
maf.File <- system.file("extdata/", "CRC_HZ.maf", package = "MesKit")
clin.File <- system.file("extdata/", "CRC_HZ.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "CRC_HZ.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
getPhyloTreeTsbLabel(phyloTree$V402)
}
