% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getSampleInfo}
\alias{getSampleInfo}
\alias{getSampleInfo,Maf-method}
\title{getSampleInfo}
\usage{
getSampleInfo(object)

\S4method{getSampleInfo}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Sample information
}
\description{
getSampleInfo
}
\examples{
getSampleInfo(maf)
}
