% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{getTMB}
\alias{getTMB}
\title{getTMB}
\usage{
getTMB(maf, tsb = c("All"), minvaf = 0.02, maxvaf = 1)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{tsb}{specify excute ITH asessment for single or all samples from the same patient(Tumor_Sample_Barcodes)}

\item{minvaf}{Default: 0. Option: on the scale of 0 to 1. Filter low frequency variants.}

\item{maxvaf}{Default: 0. Option: on the scale of 0 to 1. Filter extremely high frequency variants.}
}
\value{
a list containing MATH scores of all/selected samples.
}
\description{
Calculate tumor mutational burden(TMB) and VAF value could be filtered with minvaf and maxvaf.
}
\examples{
getTMB(maf, tsb=c("All"), minvaf=0.02, maxvaf=1)
getTMB(maf, tsb=c("SRR3670035"))

}
