% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTree}
\alias{getTree}
\alias{getTree,phyloTree-method}
\title{getTree}
\usage{
getTree(object)

\S4method{getTree}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Tree object of phyloTree
}
\description{
getTree
}
\examples{
getTree(phyloTree)
}
