% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTreeMethod}
\alias{getTreeMethod}
\alias{getTreeMethod,phyloTree-method}
\title{getTreeMethod}
\usage{
getTreeMethod(object)

\S4method{getTreeMethod}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Tree construction method of phyloTree
}
\description{
getTreeMethod
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile = maf.File, ccfFile = ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
getTreeMethod(phyloTree$HCC5647)
}
