% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_NJtree.R
\name{GO_analysis}
\alias{GO_analysis}
\title{Use R code to do the GO analysis.}
\usage{
GO_analysis(genes = NULL, GO.type = GO.type, pval = pval,
  pAdjustMethod = pAdjustMethod, qval = qval, outdir = outdir,
  patientID = patientID, name = name)
}
\arguments{
\item{GO.type}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three. Default type="ALL"}

\item{pval}{Cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default pAdjustMethod="BH"}

\item{qval}{Cutoff value of qvalue. Default qval=0.2}

\item{njtree}{a njtree object generated by NJtree.R}

\item{savePlot}{logical. Whether to print the dotplot or barplot of GO result.}

\item{writeTable}{logical. Whether to print the table of GO result.}

\item{plotType}{one of "dot" , "bar"}

\item{showCategory}{category numbers}
}
\value{
to be continued
}
\description{
Use R code to do the GO analysis.
}
\examples{
GO.njtree(njtree, GO.type = "BP", savePlot = T, writeTable = T)

}
