% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.shiny.R
\name{GO_analysis.shiny}
\alias{GO_analysis.shiny}
\title{Use R code to do the GO analysis.}
\usage{
GO_analysis.shiny(genes = NULL, GO.type = GO.type, pval = pval,
  pAdjustMethod = pAdjustMethod, qval = qval, outdir = outdir,
  patientID = patientID, Name = Name, savePlot = savePlot,
  plotType = plotType, showCategory = showCategory)
}
\arguments{
\item{GO.type}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three. Default type="ALL"}

\item{pval}{Cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default pAdjustMethod="BH"}

\item{qval}{Cutoff value of qvalue. Default qval=0.2}

\item{savePlot}{logical. Whether to print the dotplot or barplot of GO result.}

\item{plotType}{one of "dot" , "bar"}

\item{showCategory}{category numbers}

\item{njtree}{a njtree object generated by NJtree.R}
}
\description{
Use R code to do the GO analysis.
}
\examples{
GO.njtree(njtree, GO.type = "BP", savePlot = T)

}
