% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOtree.R
\name{GOTree}
\alias{GOTree}
\title{GOTree}
\usage{
GOTree(
  phyloTree,
  GO.type = "BP",
  pval = 0.05,
  pAdjustMethod = "BH",
  qval = 0.2,
  plotType = "dot",
  showCategory = 5
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getPhyloTree function}

\item{GO.type}{one of "BP", "MF", and "CC" sub-ontologies, or "ALL" for pooling 3 GO sub-ontologies. Default type="ALL"}

\item{pval}{cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default pAdjustMethod="BH"}

\item{qval}{cutoff value of qvalue. Default qval=0.2}

\item{plotType}{one of "dot", "bar"}

\item{showCategory}{category numbers}
}
\value{
GO enrichment results
}
\description{
Perform GO enrichment analysis of trunk/branch mutations of a phylogenetic tree.
}
\examples{
GOTree(phyloTree, GO.type = "BP")
}
