% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJtree_heatmap.R
\name{heatmap_input}
\alias{heatmap_input}
\title{plot the corresponding binary heatmap| ccf heatmaf of the NJtree}
\usage{
heatmap_input(maf, use.indel, use.ccf, ccf, ccf.mutation.id,
  ccf.mutation.sep)
}
\arguments{
\item{use.indel}{logical. Whether to uses all variants irrespective of values in Mutation_Status.Defaults to FALSE, only use somatic mutations.}

\item{use.ccf}{logical. Whether to print the binary heatmap with ccf values. Defaults to FALSE, print the binary heatmap without ccf value}

\item{ccf.mutation.id}{manually specify which columns could be joint by ccf.mutation.sep to get the same format of mutation id in ccfy}

\item{ccf.mutation.sep}{manually specify the separator character.Values on each line of the ccf.mutation.id are separated by this character. I}

\item{is.ccf}{Use mutation CCF for heatmap painting. Default FALSE.}
}
\value{
binary mutation heatmap or mutation CCF heatmap
}
\description{
plot the corresponding binary heatmap| ccf heatmaf of the NJtree
}
\examples{
mut.heatmap(Maf, use.indel = F)
mut.heatmap(Maf, use.indel = F, use.ccf = TRUE)
}
