% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{JSI_dist}
\alias{JSI_dist}
\title{calJSI}
\usage{
JSI_dist(df)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{plot}{logical, whether to show the plot, default TRUE}

\item{use.circle, }{whether to use circle as background of the correlation matrix plot, 
default TURE}

\item{title}{the title of the plot, default is "Jaccard similarity"}
}
\value{
correlation matrix and heatmap via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity coefficient is defined as the ratio of shared variants to all
(shared plus discordant) variants for two samples.
}
\examples{
calJSI(maf)
}
