% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_sort.R
\name{maf_preprocess}
\alias{maf_preprocess}
\title{get the sorted binary| ccf mutation matrix}
\usage{
maf_preprocess(maf.dat, use.ccf = F, ccf, ccf.mutation.id, ccf.mutation.sep)
}
\description{
get the sorted binary| ccf mutation matrix
}
