% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/classS4.R
\docType{class}
\name{MafList-class}
\alias{MafList-class}
\alias{MafList}
\title{MafList}
\description{
S4 class for storing multi-patients Maf object.

S4 class for storing multi-patients Maf object.
}
