% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH_Score.R
\name{MATH_score}
\alias{MATH_score}
\title{A Multifunctional MATH Calculator}
\usage{
MATH_score(maf, tsb = c("OFA"), minvaf = 0, maxvaf = 1)
}
\arguments{
\item{tsb}{specify single/all sample names (Tumor_Sample_Barcodes)}

\item{minvaf}{filter low frequency variants caused by sequencing error. Default 0. (on the scale of 0 to 1)}

\item{maxvaf}{filter high frequency variants due to copy number alterations or impure tumor. Default 1. (on the scale of 0 to 1)}

\item{maf_file}{specify a maf document/directory as the input of the function}
}
\value{
MATH score list for all/selected samples or a single MATH score value
}
\description{
read maf file as data.frame, calculate MATH score and present results in different ways determined by parameters.
This function requires VAF for MATH score calculation and Tumor_sample_Barcode for selection.
VAF can be on the scale 0-1 and you could filter VAF with minvaf and maxvaf.
}
\examples{
\dontrun{
maf_file = "/home/ninomoriaty/R_Project/311252_snv_indel.imputed.maf"
MATH_score(maf_file)
MATH_score(maf_file, tsb = c("tsb1"))
MATH_score(maf_file, tsb = c("tsb1", "tsb2", "tsb3"))
}
}
