% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{mathScore}
\alias{mathScore}
\title{mathScore}
\usage{
mathScore(maf, patient.id = NULL, withinTumor = FALSE, min.vaf = 0.02, ...)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{withinTumor}{Calculate AUC within types in each patients. Default :FALSE.}

\item{min.vaf}{The minimum VAF for filtering variants. Default: 0.02}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\value{
A data.frame of MATH scores
}
\description{
calculates MATH score of each tumor sample or based on Mutant-Allele Tumor Heterogeneity (MATH) approach.
}
\examples{
mathScore(maf)
}
\references{
Mroz, Edmund A. et al. Intra-Tumor Genetic Heterogeneity and Mortality in Head and Neck Cancer: Analysis of Data from The Cancer Genome Atlas. Ed. Andrew H. Beck. PLoS Medicine 12.2 (2015): e1001786.
}
