% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meskit.R
\docType{package}
\name{Meskit-package}
\alias{Meskit}
\alias{Meskit-package}
\title{Meskit: Analyse and visualize multi-region whole-exome sequencing data}
\description{
Analysis and visualize multi-region whole-exome sequencing data.This package provides various functions to perform most commonly used analyses in cancer genomics and to create feature rich customizable visualzations with minimal effort.
}
\keyword{internal}
