% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutSharedPrivate.R
\name{mutSharedPrivate}
\alias{mutSharedPrivate}
\title{Use R code to find the intersect mutations and their types in several samples of one patient}
\usage{
mutSharedPrivate(maf, show.num = FALSE)
}
\arguments{
\item{maf}{Maf object return from read.Maf()}

\item{show.num}{a logic parameter to determine whether to show the number of each mutations in the stack plot}
}
\value{
mutSharedPrivate
}
\description{
Use R code to find the intersect mutations and their types in several samples of one patient
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "Meskit")
maf <- readMaf(patientID = "311252", mafFile = maf.File, sampleInfo = sampleInfo.File, refBuild = "hg19")
mutSharedPrivate(maf)
}
