% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R, R/mutSigSummary.R
\name{mutSigSummary}
\alias{mutSigSummary}
\title{mutSigSummary}
\usage{
mutSigSummary(treeMSOutput)

mutSigSummary(treeMSOutput)
}
\arguments{
\item{treeMSOutput}{The output of function treeMutationalSig.}

\item{treeMSOutput}{The output of function treeMutationalSig.}
}
\value{
Data frame of each set/branch's mutational signature.

Data frame of each set/branch's mutational signature.
}
\description{
Provide a summary data frame for signatures in different branches.

Provide a summary data frame for signatures in different branches.
}
\examples{
signature1 <- sumMutationalSig(treeMSOutput)

mutSig.summary <- mutSigSummary(treeMSOutput)

}
