% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackPlot.R
\name{mutStackPlot}
\alias{mutStackPlot}
\title{Stack Plot for variatns of oncogenes and Tumor Suppression Gene(TSGs) in patient}
\usage{
mutStackPlot(maf, oncogeneListFile, tsgListFile, show.percentage = FALSE)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{oncogeneListFile}{the file containing all oncogenes needed to be checked.}

\item{tsgListFile}{the file containing all TSGs needed to be checked.}

\item{show.percentage}{logical parameter that determine whether the 
percentage of different intersections will print in the stack plot or not}

\item{themeOption}{select a theme Palette from ggsci. Default "aaas".}
}
\description{
With the gene list passed by oncogeneListFile and tsgListFile, 
mutStackPlot can calculate the quantity of variants from different 
intersections and summarize the result as a stack plot. The stacks will be 
constructed by three part: 1. Private: variants specific for each samples;
2. Parital-shared: variants shared by multiple samples but not existing in 
all samples; 3. Shared: variants shared by all samples.
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
oncogeneListFile <- system.file("extdata/", "oncogene.list.txt", package = "Meskit")
tsgListFile <- system.file("extdata/", "TSG.list.txt", package = "Meskit")
mutStackPlot(maf, oncogeneListFile, tsgListFile, themeOption="npg", show.percentage=TRUE)

}
