% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTrunkBranch.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(
  phyloTree,
  CT = FALSE,
  geneList = NULL,
  pvalue = 0.05,
  plot = TRUE,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree}{object generated by phyloTree function.}

\item{CT}{Distinction between C>T at CpG and C>T at other sites, Default FALSE}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{pvalue}{confidence level of the interval for Fisher test. Default: 0.05.}

\item{plot}{output a list of diagrams that visualize mutational signature of trunk/branches in a phylogenetic tree.Default FALSE.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}
}
\value{
a list of box plots based on mutational categories
}
\description{
Summarize and conduct paired Wilcoxon test of mutations of trunk/branches in a phylogenetic tree.
}
\examples{
mutTrunkBranch(phyloTree)
mutTrunkBranch(phyloTree,plot = T)

}
