% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMut.R
\name{mutType_SP}
\alias{mutType_SP}
\title{classifyMut}
\usage{
mutType_SP(
  Tumor_Type,
  unique_barcode_count,
  total_barcode_count,
  type_barcode_count,
  TypeCount,
  classByType
)
}
\arguments{
\item{classByType}{FALSE(Default). Define shared pattern of mutations based on tumor types (TRUE) or samples (FALSE)}

\item{maf}{the Maf object generated by readMaf
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{class}{the class which would be represented, default is "SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS".}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{plot}{TRUE(Default).}

\item{topGenesCount}{the number of genes print, default is 10}

\item{remove_empty_columns}{whether remove the samples without alterations. Only works when plot is TRUE}

\item{remove_empty_rows}{whether remove the genes without alterations. Only works when plot is TRUE}

\item{showColnames}{TRUE(Default). Show sample names of columns.}

\item{patientsCol}{a list containing customized colors for distinct patients. Default: NULL.}

\item{bgCol}{background grid color. Default: "#f0f0f0"}
}
\description{
classifyMut
}
\examples{
classifyMut(maf, class = "SP")
}
