% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNeiDist.R
\name{nei_dist}
\alias{nei_dist}
\title{calFst}
\usage{
nei_dist(df)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.VAF}{specify the minimum CCF value of subclonal mutations}

\item{plot}{logical, default TRUE}

\item{use.circle, }{default FALSE}
}
\value{
Nei's genetic distance matrix of sample-pairs from the same patient
}
\description{
Nei's distance of CCF for each sample-pair
}
\examples{
calNeiDist(maf)
}
\references{
Lee JK, Wang J, Sa JK, et al. Spatiotemporal genomic architecture informs precision oncology in glioblastoma. 
Nat Genet. 2017;49(4):594–599. doi:10.1038/ng.3806
}
