% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJtree.R
\name{NJtree}
\alias{NJtree}
\title{Class NJtree}
\usage{
NJtree(maf, use.indel = FALSE, use.ccf = FALSE, mut.signature = TRUE,
  sig.min.mut.number = 50, ccf.mutation.id, ccf.mutation.sep)
}
\arguments{
\item{patientID}{patient name}
}
\value{
NJtree object
}
\description{
Class NJtree
}
\examples{
maf <- read.Maf(patientID)
njtree <- read.NJtree(maf)
getNJtreenj(njtree)
getNJtreemut_sort(njtree)
getNJtreeSignature(njtree)
}
