% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway.shiny.R
\name{Pathway_analysis.shiny}
\alias{Pathway_analysis.shiny}
\title{USE R code to do the pathway analysis.}
\usage{
Pathway_analysis.shiny(genes = NULL, pathway.type = pathway.type,
  pval = pval, pAdjustMethod = "BH", qval = qval, outdir = NULL,
  patientID = patientID, Name = Name, savePlot = TRUE,
  plotType = plotType, showCategory = showCategory)
}
\arguments{
\item{pathway.type}{one of "KEGG" or "Reactome". Default type = "KEGG"}

\item{pval}{Cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default pAdjustMethod="BH"}

\item{qval}{Cutoff value of qvalue. Default qval=0.2}

\item{savePlot}{logical. Whether to print the dotplot or barplot of GO result.}

\item{plotType}{one of "dot" , "bar"}

\item{showCategory}{category numbers}

\item{njtree}{a njtree object generated by NJtree.R}
}
\description{
USE R code to do the pathway analysis.
}
\examples{
Pathway.njtree(njtree, pathway.type = "KEGG", savePlot = T)
}
