% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayTree.R
\name{pathwayTree}
\alias{pathwayTree}
\title{pathwayTree}
\usage{
pathwayTree(
  phyloTree,
  pathway.type = "KEGG",
  pval = 0.05,
  pAdjustMethod = "BH",
  qval = 0.2,
  plotType = "dot",
  showCategory = 5
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getPhyloTree function.}

\item{pathway.type}{one of "KEGG" or "Reactome". Default type="KEGG"}

\item{pval}{cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default pAdjustMethod="BH"}

\item{qval}{cutoff value of qvalue. Default qval=0.2}

\item{plotType}{one of "dot", "bar"}

\item{showCategory}{category numbers}
}
\value{
pathway enrichment results
}
\description{
Perform pathway enrichment analysis of trunk/branch mutations of a phylogenetic tree.
}
\examples{
pathwayTree(phyloTree, pathway.type = "KEGG")
}
