% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PhyloTree.R
\name{phylotreeInput}
\alias{phylotreeInput}
\title{A object of NJtree painter}
\usage{
phylotreeInput(Phylo, signature = "", show.mutSig, phylotree.type)
}
\arguments{
\item{maf}{MAF object}
}
\value{
NJtree plot (phylogenetic tree and heatmap)
}
\description{
A object of NJtree painter
}
\examples{
maf <- read.Maf("311252",dat.dir = './data/multi_lesion', BSG = "BSgenome.Hsapiens.UCSC.hg19")
plotPhyloTree(maf)
}
