% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/classS4.R
\docType{class}
\name{phyloTreeList-class}
\alias{phyloTreeList-class}
\title{phyloTreeList class}
\description{
S4 class for storing multi-patients phyloTree.

S4 class for storing multi-patients phyloTree.
}
