% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{plotMutationalSig}
\alias{plotMutationalSig}
\title{plotMutationalSig}
\usage{
plotMutationalSig(treeMSOutput)
}
\arguments{
\item{treeMSOutput}{The output of function treeMutationalSig.}
}
\value{
A signature plot shows the signature for each branch and presents the distribution in each mutational categories.
}
\description{
Return a matrix of panels based on mutational categories
}
\examples{
plotMutationalSig(treeMSOutput)

}
