% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutProfile.R
\name{plotMutProfile}
\alias{plotMutProfile}
\title{plotMutProfile}
\usage{
plotMutProfile(
  maf,
  patient.id = NULL,
  class = "SP",
  classByTumor = FALSE,
  topGenesCount = 10,
  geneList = NULL,
  bgCol = "#f0f0f0",
  patientsCol = NULL,
  remove_empty_columns = TRUE,
  remove_empty_rows = TRUE,
  showColnames = TRUE
)
}
\arguments{
\item{maf}{The Maf object generated by readMaf}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{class}{The class which would be represented, default is "SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS".}

\item{classByTumor}{FALSE(Default). Define shared pattern of mutations based on tumor types (TRUE) or samples (FALSE)}

\item{topGenesCount}{The number of genes print, default is 10}

\item{geneList}{A list of genes to restrict the analysis. Default NULL.}

\item{bgCol}{Background grid color. Default: "#f0f0f0"}

\item{patientsCol}{A list containing customized colors for distinct patients. Default: NULL.}

\item{remove_empty_columns}{Whether remove the samples without alterations. Only works when plot is TRUE}

\item{remove_empty_rows}{Whether remove the genes without alterations. Only works when plot is TRUE}

\item{showColnames}{TRUE(Default). Show sample names of columns.}
}
\description{
plotMutProfile
}
\examples{
plotMutProfile(maf, class = "SP")
}
