% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationalSig.R
\name{plotMutSig}
\alias{plotMutSig}
\title{plotMutSig}
\usage{
plotMutSig(tree.mutSig)
}
\arguments{
\item{tree.mutSig}{The output of function treeMutationalSig.}
}
\value{
A signature plot shows the signature for each branch and presents the distribution in each mutational categories.
}
\description{
Visualize the mutational signatures of trunk/branches in a phylogenetic tree.
}
\examples{
plotMutSig(tree.mutSig)

}
