% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfile}
\alias{plotMutSigProfile}
\title{plotMutSigProfile}
\usage{
plotMutSigProfile(fit_out, patient.id = NULL, mode = NULL)
}
\arguments{
\item{fit_out}{result generated by fitSignatures function.}

\item{patient.id}{patient.id select the specific patients. Default: NULL, all patients are included.}

\item{mode}{type of mutation spectrum.Default: NULL. Options:'Original','Reconstructed' or 'Difference'}
}
\value{
mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
tri_matrix <- triMatrix(phyloTree)
fit_out <- fitSignatrues(tri_matrix)
plotMutSigProfile(fit_out)
}
