% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfiler.R
\name{plotMutSigProfiler}
\alias{plotMutSigProfiler}
\title{plotMutSigProfiler}
\usage{
plotMutSigProfiler(tree.mutSig, patient.id = NULL)
}
\arguments{
\item{tree.mutSig}{object generated by treeMutSig function.}

\item{patient.id}{patient.id select the specific patients. Default: NULL, all patients are included.}
}
\value{
a list of mutation signature profile of patients
}
\description{
plotMutSigProfiler
}
\examples{
tree.mutSig <- treeMutSig(phyloTree)
plotMutSigProfiler(tree.mutSig)
}
