% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrunkBranch.R
\name{plotTrunkBranch}
\alias{plotTrunkBranch}
\title{plotTrunkBranch}
\usage{
plotTrunkBranch(tree.mutSig, conf.level = 0.95)
}
\arguments{
<<<<<<< HEAD
\item{tree.mutSig}{the output of function treeMutationalSig.}
=======
\item{treeMSOutput}{the output of function treeMutSig.}
>>>>>>> 3a9fd47c60a294dec42f135a96cdc9898bedba7a

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95.}
}
\value{
Box plots based on mutational categories
}
\description{
Prints the distribution of branch/trunk mutations based on mutational categories.
}
\examples{
plotTrunkBranch(tree.mutSig, conf.level = 0.95)
}
