% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAF_plot_alpha.R
\name{plotVAF}
\alias{plotVAF}
\title{A VAF plot painter}
\usage{
plotVAF(maf, sampleOption = "OFA", themeOption = "aaas",
  fileFormat = "png", showMATH = TRUE)
}
\arguments{
\item{sampleOption}{specify single/all sample names 
(Tumor_Sample_Barcodes, tsb). Default "OFA".}

\item{themeOption}{select a coloring scheme from ggsci. 
Default "aaas".}

\item{fileFormat}{choose an output file format accessable for ggsave. 
Default "png".}

\item{maf_file}{specify a maf document/directory as the input of the 
function}
}
\value{
Images of selected samples' VAF
}
\description{
Read maf file as data.frame, cluster data with maftools
 and finally draw variant allele frequency(VAF) frequency distribution
 curve with ggplot2 as well as ggridges. We could use different parameters
 to control output images from different samples or conclude all samples' 
 VAF information in one image.
}
\examples{
\dontrun{
VAF_plot(maf_file, sampleOption="OFA", themeOption="aaas") 
VAF_plot(maf_file, sampleOption="All") 
VAF_plot(maf_file, sampleOption="MIX") 
VAF_plot(maf_file, sampleOption="tsb1", fileFormat="pdf") 

}
}
