% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schism2Fishplot.R
\name{prepareSchismInput}
\alias{prepareSchismInput}
\title{Prepare inputs for SCHISM}
\usage{
prepareSchismInput(dirClusterTsv, dirLociTsv, dirOutput)
}
\arguments{
\item{dir.cluster.tsv}{specify the directory of cluster.tsv document}

\item{dir.loci.tsv}{specify the directory of loci.tsv document}

\item{dir.output}{specify the directory of two output for SCHISM}
}
\value{
clusterEstimates.tsv, mutation_to_cluster.tsv
}
\description{
read tsv documents from PyClone and figure 
out the list of filtered clusters. According to those 
targeted clusters, filter targeted mutations and reorganize 
data as two outputs: clusterEstimates.tsv and 
mutation_to_cluster.tsv, which are the input of SCHISM.
}
\examples{
\dontrun{
prepareSchismInput(dir.cluster.tsv, dir.loci.tsv, dir.output)
}
}
