% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNJtree.R
\name{read.njtree}
\alias{read.njtree}
\title{Get sampleID of private branchs and shared branches/trunk in NJtree}
\usage{
read.njtree(mat.nj)
}
\arguments{
\item{mat.nj}{a nj tree object generated from ape}
}
\value{
a dataframe with sampleID of each branch and trunk
}
\description{
Get sampleID of private branchs and shared branches/trunk in NJtree
}
