% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhyloTree.R
\name{readPhyloTree}
\alias{readPhyloTree}
\title{readPhyloTree}
\usage{
readPhyloTree(tree)
}
\arguments{
\item{tree}{a phylo tree object}
}
\value{
a sample IDs of trunk/branches of the phylogenetic tree
}
\description{
readPhyloTree
}
