% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meskit.R
\name{runMesKit}
\alias{runMesKit}
\title{Run the default MesKit app for analysis locally}
\usage{
runMesKit()
}
\value{
a shiny app window
}
\description{
\code{runMesKit} run MesKit locally
}
\examples{
\donttest{
runMesKit()
}
}
\author{
Mengni Liu
}
