% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schism2Fishplot.R
\name{schism2Fishplot}
\alias{schism2Fishplot}
\title{Get outputs from SCHISM and draw the fishplot}
\usage{
schism2Fishplot(dirClusterCellularity, dirGAconsensusTree)
}
\arguments{
\item{dir.cluster.cellularity}{specify the directory of cluster.cellularity 
document}

\item{dir.GA.consensusTree}{specify the directory of GA.consensusTree 
document}
}
\value{
the fishplot showing the evolution relationship predicted by SCHISM
}
\description{
read cluster.cellularity and GA.consensusTree documents 
from SCHISM. Construct fraction table and parents vector for drawing 
fishplot.
}
\examples{
\dontrun{
schism2Fishplot(dir.cluster.cellularity, dir.GA.consensusTree)
}
}
