% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMaf.R
\name{subsetMaf}
\alias{subsetMaf}
\title{Subset Maf object}
\usage{
subsetMaf(
  maf,
  geneList = NULL,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.vaf = 0,
  max.vaf = 1,
  min.average.vaf = 0,
  min.average.adj.vaf = 0,
  min.ccf = 0,
  min.ref.depth = 0,
  min.alt.depth = 0,
  min.total.depth = 0,
  clonalStatus = NULL
)
}
\arguments{
\item{geneList}{subset by geneList.Default:NULL}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSilent".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.vaf}{The minimum VAF for filtering variants. Default: 0.}

\item{max.vaf}{The maximum VAF for filtering variants. Default: 1.}

\item{min.average.vaf}{The minimum tumor average VAF for filtering variants. Default: 0.}

\item{min.average.adj.vaf}{The minimum tumor average ajust VAF for filtering variants. Default: 0.}

\item{min.ccf}{The minimum CCF for filtering variants. Default: 0.}

\item{min.ref.depth}{The minimum reference allele depth for filtering variants. Default: 0.}

\item{min.alt.depth}{The minimum alteratation allele depth for filtering variants. Default: 0.}

\item{min.total.depth}{The minimum total allele depth for filtering variants. Default: 0.}

\item{clonalStatus}{subset by clonal status.Default: NULL.Option: "Clonal","Subclonal".}
}
\value{
Maf object.
}
\description{
Read tab delimited MAF (can be plain text or *.gz compressed) file along with sample information file.
}
\examples{
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
maf <- subsetMaf(maf)
}
