% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNeutral.R
\name{testNeutral}
\alias{testNeutral}
\title{testNeutral}
\usage{
testNeutral(
  maf,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  withinTumor = FALSE,
  use.adjVAF = FALSE,
  min.depth = 2,
  min.vaf = 0.1,
  max.vaf = 0.3,
  R2.threshold = 0.98,
  min.mut.count = 20,
  plot = TRUE
)
}
\arguments{
\item{maf}{The Maf object generated by readMaf}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, chrX, chrY. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{use.adjVAF}{Whether to use the adjusted VAF in analysis. Default: FALSE.}

\item{min.depth}{The minimun total depth of coverage. Defalut: 2}

\item{min.vaf}{The minimum value of adjusted VAF value. Default: 0.1}

\item{max.vaf}{The maximum value of adjusted VAF value. Default: 0.3}

\item{R2.threshold}{The threshod of R2 to decide whether a tumor follows neutral evolution. Default: 0.98}

\item{min.mut.count}{The minimun number of subclonal mutations used to fit model. Default: 20}

\item{plot}{Logical, whether to print model fitting plot of each sample. Default: TRUE}
}
\description{
Evaluate whether a tumor follows neutral evolution or under strong selection during the growth based on variant frequency distribution (VAF) of subclonal mutations.
The subclonal mutant allele frequencies of a follow a simple power-law distribution predicted by neutral growth.
}
\examples{
testNeutral(maf)
}
\references{
Williams, M., Werner, B. et al. Identification of neutral tumor evolution across cancer types. Nat Genet 48, 238-244 (2016)
}
