% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNeutral.R
\name{testPowerLaw}
\alias{testPowerLaw}
\title{testNeutral}
\usage{
testPowerLaw(df, min.VAF_adj, max.VAF_adj, min.depth, R2.threshold, plot)
}
\arguments{
\item{min.depth}{the minimun depth of coverage. Defalut: 10}

\item{R2.threshold}{a threshod of R2 to decide whether a tumor follows neutral evolution. Default: 0.98}

\item{plot, }{logical. Whether print model fitting plot of each sample. TRUE (Default)}

\item{maf}{the Maf object generated by readMaf}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\description{
Evaluat whether a tumor follows neutral evolution or under strong selection
during the growth based on variant frequency distribution (VAF) of subclonal 
The subclonal mutant allele frequencies of a follow a simple power-law distribution predicted by neutral growth.
}
\examples{
testNeutral(maf)
}
\references{
Williams MJ, Werner B, Barnes CP, Graham TA, Sottoriva A. Identification of neutral tumor evolution across cancer types. 
Nat Genet. 2016;48(3):238–244. doi:10.1038/ng.3489
}
