% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeGO.R
\name{treeGO}
\alias{treeGO}
\title{treeGO}
\usage{
treeGO(
  phyloTree,
  driverGenesFile = NULL,
  GO.type = "BP",
  pval = 0.05,
  pAdjustMethod = "BH",
  qval = 0.2,
  plotType = "dot",
  showCategory = 5,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getPhyloTree function}

\item{driverGenesFile}{The file with driver gene list. Default NULL.}

\item{GO.type}{one of "BP", "MF", and "CC" sub-ontologies, 
or "ALL" for pooling 3 GO sub-ontologies. Default is "ALL"}

\item{pval}{cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
Default is "BH"}

\item{qval}{cutoff value of qvalue. Default qval=0.2}

\item{plotType}{one of "dot", "bar", default is "dot"}

\item{showCategory}{number of categories will be shown, default is 5}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{
GO enrichment results
}
\description{
Perform GO enrichment analysis of trunk/branch mutations of a phylogenetic tree.
}
\examples{
treeGO(phyloTree, GO.type = "BP")
}
