% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutationalSig}
\alias{treeMutationalSig}
\title{Check Mutational Signature for each branch of phylogenetic tree}
\usage{
treeMutationalSig(mafData, branch, patientID, refBuild,
  driverGenesDir = FALSE, mutThreshold = 50)
}
\arguments{
\item{maf_file}{specify a maf document/directory as the input of the 
function}

\item{branch_file}{specify a txt document/directory as the input of 
the branches (needed to be refined)}
}
\value{
data frame of each set/branch's mutational signature.
}
\description{
Read maf file as data.frame. Define branches' set relationship
 by re-labeling their tumor sample barcode from the smallest set. Calcualte 
 each branch's mutational signature weight according to cosmic reference and
 pick the maxium. Return a data frame of each set/branch's mutational 
 signature.
}
\examples{
\dontrun{
treeMutationalSig(maf_file, branch_file)
treeMutationalSig(maf_file, branch_file, driver_genes_dir)
treeMutationalSig(maf_file, branch_file, 
driver_genes_dir, mut.threshold=30)
}
}
