% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutSig}
\alias{treeMutSig}
\title{treeMutSig}
\usage{
treeMutSig(
  phyloTree,
  driverGenesFile = NULL,
  min.mut.count = 15,
  signaturesRef = "cosmic",
  plot = FALSE
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by phyloTree function.}

\item{driverGenesFile}{The file with driver gene list. Default NULL.}

\item{min.mut.count}{the threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic". Option: "nature2013".}

\item{plot}{output a list of diagrams that visualize mutational signature of trunk/branches in a phylogenetic tree.Default FALSE.}
}
\value{
a list of data frames, each one contains treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in phyloTree object and output the treeMSOutput for further summary and plot functions.
}
\examples{
treeMutSig(phyloTree, driverGenesFile=NULL, min.mut.count=15, signaturesRef="cosmic")
treeMutSig(phyloTree, plot = T)
}
