% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumorCloneCCF.R
\name{tumorClonesPlot}
\alias{tumorClonesPlot}
\title{tumorClonesPlot}
\usage{
tumorClonesPlot(maf, clone.min.mut = 5, clone.min.aveCCF = 0.1)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{clone.min.mut}{specify the minimum mutation number of a subclone.}

\item{clone.min.aveCCF}{specify the minimum average CCF value of a subclone.}
}
\value{
CCF distribution of subclones and mutations in the form of radar plot and dot plot, respectively.
}
\description{
Visualize the CCF distribution of subclones and mutations((loci.tsv and cluster.tsv generated by PyClone are required).
}
\examples{
tumorClonesPlot(maf)
importFrom ggridges scale_discrete_manual
}
