% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAF_plot_alpha.R
\name{VAF_plot}
\alias{VAF_plot}
\title{A VAF plot painter}
\usage{
VAF_plot(maf, sample_option = "OFA", theme_option = "aaas",
  file_format = "png", show.MATH = T)
}
\arguments{
\item{sample_option}{specify single/all sample names (Tumor_Sample_Barcodes, tsb). Default "OFA".}

\item{theme_option}{select a coloring scheme from ggsci. Default "aaas".}

\item{file_format}{choose an output file format accessable for ggsave. Default "png".}

\item{maf_file}{specify a maf document/directory as the input of the function}
}
\value{
Images of selected samples' VAF
}
\description{
Read maf file as data.frame, cluster data with maftools and finally draw variant allele frequency(VAF)
frequency distribution curve with ggplot2 as well as ggridges. We could use different parameters to control output 
images from different samples or conclude all samples' VAF information in one image.
}
\examples{
\dontrun{
VAF_plot(maf_file, sample_option = "OFA", theme_option = "aaas") # draw a VAF image that contains all samples' VAF distribution curves with different themes.
VAF_plot(maf_file, sample_option = "All") # draw VAF images for every sample respectively
VAF_plot(maf_file, sample_option = "tsb1", file_format = "pdf") # draw a VAF image for sample tsb1 and save as a pdf file.

}
}
