% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFst.R
\name{calFst}
\alias{calFst}
\title{calFst}
\usage{
calFst(
  maf,
  patient.id = NULL,
  min.vaf = 0.02,
  min.total.depth = 2,
  plot = TRUE,
  withinTumor = FALSE,
  use.circle = TRUE,
  title = NULL,
  number.cex = 8,
  number.col = "#C77960",
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{min.vaf}{Specify the minimum VAF_adj, default is 0.02}

\item{min.total.depth}{The minimum total allele depth for filtering variants. Default: 2.}

\item{plot}{Logical (Default:TRUE). Whether to show the plot, default TRUE}

\item{withinTumor}{Calculate fst within types in each patients,default is FALSE.}

\item{use.circle}{Logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix}

\item{title}{The title of the plot. Default is "Nei's distance"}

\item{number.cex}{The size of text shown in correlation plot. Default 8.}

\item{number.col}{The color of text shown in correlation plot. Default "#C77960".}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\value{
A list contains Fst value of MRS and Hudson estimator of each sample-pair, respectively.
}
\description{
Genetic divergence between regions of subclonal sSNVs using the Weir and Cockerham method
}
\examples{
calFst(maf)
}
\references{
Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. Nat Genet. 2017;49(7):1015-1024.

Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and interpreting FST: the impact of rare variants. Genomic Res. 2013;23(9):1514-1521.
}
