% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFst.R
\name{calFst}
\alias{calFst}
\title{calFst}
\usage{
calFst(
  maf,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.vaf = 0.02,
  min.total.depth = 2,
  plot = TRUE,
  use.circle = TRUE,
  title = NULL,
  withinTumor = FALSE,
  number.cex = 8,
  number.col = "#C77960"
)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSilent".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.vaf}{specify the minimum VAF_adj, default is 0.02}

\item{min.total.depth}{The minimum total allele depth for filtering variants. Default: 2.}

\item{withinTumor}{calculate fst within types in each patients,default is FALSE.}
}
\value{
a list contains Fst value of MRS and Hudson estimator of each sample-pair, respectively.
}
\description{
genetic divergence between regions of subclonal sSNVs using the Weir and Cockerham method
}
\examples{
calFst(maf)
}
\references{
Sun R, Hu Z, Sottoriva A, et al. Between-region genetic divergence reflects the mode and tempo of tumor evolution. Nat Genet. 2017;49(7):1015–1024. doi:10.1038/ng.3891

Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and interpreting FST: the impact of rare variants. 
Genome Res. 2013;23(9):1514–1521. doi:10.1101/gr.154831.113
}
