% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{calJSI}
\alias{calJSI}
\title{calJSI}
\usage{
calJSI(maf, type = "full", plot = TRUE)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function}

\item{type}{character, "full" (default), "upper" or "lower", display full matrix, lower triangular or upper triangular matrix.}

\item{plot}{logical, default TRUE}
}
\value{
the graphical display of correlation matrix via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity coefficient is defined as the ratio of shared variants to all 
(shared plus discordant) variants for two samples.
}
\examples{
calJSI(maf)
}
