% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{calMATH}
\alias{calMATH}
\title{mathScore}
\usage{
calMATH(VAF)
}
\arguments{
\item{maf}{a Maf object generated by readMaf.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.vaf}{the minimum VAF for filtering variants. Default: 0.02}

\item{plot}{TRUR(default)}
}
\value{
a list containing MATH scores and violin plot of all samples
}
\description{
Calculate MATH score and VAF value could be filtered with minvaf and maxvaf.
}
\examples{
mathScore(maf, min.vaf=0.02, max.vaf=1)
}
