% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfAUC.R
\name{ccfAUC}
\alias{ccfAUC}
\title{ccfAUC}
\usage{
ccfAUC(maf, tsb = NULL, auc.cutoff = 0.145)
}
\arguments{
\item{maf}{a Maf object generated by readMaf.}

\item{tsb}{specify Tumor_Sample_Barcodes for single or all samples from maf of the patient. Default: NULL, which means that output all samples' AUC Score.}
}
\value{
A list containing AUC scores' data frame and corresponding ccf/proporation plot.
}
\description{
Calculate AUC score and provide a ccf/proporation plot.
}
\examples{
ccfAUC(maf, tsb=NULL)
ccfAUC(maf, tsb=c("SRR3670035"))

}
