% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfAUC.R
\name{ccfAUC}
\alias{ccfAUC}
\title{ccfAUC}
\usage{
ccfAUC(
  maf,
  patient.id = NULL,
  min.ccf = 0,
  withinTumor = FALSE,
  plot.density = TRUE,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{min.ccf}{The minimum value of CCF. Default: 0}

\item{withinTumor}{Calculate AUC within types in each patients,default is FALSE.}

\item{plot.density}{Whether to show the density plot. Default: TRUE}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\value{
A list containing AUC of CCF and a graph
}
\description{
The tumor heterogeneity was estimated as the area under the curve (AUC) of the cumulative density function from all cancer cell fractions per tumor
}
\examples{
ccfAUC(maf)

}
\references{
Charoentong P, Finotello F, et al. Pan-cancer Immunogenomic Analyses Reveal Genotype-Immunophenotype Relationships and Predictors of Response to Checkpoint Blockade. Cell reports 2017, 18:248-262.
}
