% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMut.R
\name{classifyMut}
\alias{classifyMut}
\title{classifyMut}
\usage{
classifyMut(
  maf,
  class = "SP",
  patient.id = NULL,
  classByType = FALSE,
  geneList = NULL,
  plot = TRUE,
  topGenesCount = 15,
  remove_empty_columns = TRUE,
  remove_empty_rows = TRUE,
  showColnames = TRUE,
  patientsCol = NULL,
  bgCol = "#f0f0f0"
)
}
\arguments{
\item{maf}{The Maf object generated by readMaf
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{class}{The class which would be represented, default is "SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS".}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{classByType}{FALSE(Default). Define shared pattern of mutations based on tumor types (TRUE) or samples (FALSE)}

\item{geneList}{A List of genes to restrict the analysis. Default NULL.}

\item{plot}{TRUE(Default).}

\item{topGenesCount}{The number of genes print, default is 10}

\item{remove_empty_columns}{Whether remove the samples without alterations. Only works when plot is TRUE}

\item{remove_empty_rows}{Whether remove the genes without alterations. Only works when plot is TRUE}

\item{showColnames}{TRUE(Default). Show sample names of columns.}

\item{patientsCol}{A list containing customized colors for distinct patients. Default: NULL.}

\item{bgCol}{Background grid color. Default: "#f0f0f0"}
}
\value{
a data.frame with classification of mutations
}
\description{
classifyMut
}
\examples{
classifyMut(maf, class = "SP")
}
