% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMut.R
\name{classifyMut}
\alias{classifyMut}
\title{classifyMut}
\usage{
classifyMut(maf, patient.id = NULL, class = "SP", classByTumor = FALSE, ...)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{class}{The class which would be represented, default is "SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS".}

\item{classByTumor}{FALSE(Default). Classify mutations based on "Tumor_ID".}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\value{
A data.frame with classification of mutations
}
\description{
classifyMut
}
\examples{
classifyMut(maf, class = "SP")
}
