% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMut.R
\name{classifyMut}
\alias{classifyMut}
\title{classifyMut}
\usage{
classifyMut(maf, patient.id = NULL, class = "SP", classByType = FALSE)
}
\arguments{
\item{maf}{The Maf object generated by readMaf
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{class}{The class which would be represented, default is "SP" (Shared pattern: Public/Shared/Private),
other options: "CS" (Clonal status: Clonal/Subclonl) and "SPCS".}

\item{classByType}{FALSE(Default). Define shared pattern of mutations based on tumor types (TRUE) or samples (FALSE)}
}
\value{
a data.frame with classification of mutations
}
\description{
classifyMut
}
\examples{
classifyMut(maf, class = "SP")
}
