% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\docType{class}
\name{classMaf-class}
\alias{classMaf-class}
\alias{classMaf}
\title{MAF/CCF/Sample_info Reader}
\arguments{
\item{patientID}{patientID for all samples in mafFile.}

\item{mafFile}{MAF file directory.}

\item{sampleInfoFile}{sample_info.txt file directory.}

\item{ccfClusterFile}{CCF cluster.tsv file directory if ccf data provided. Default NULL.}

\item{ccfLociTsvFile}{CCF loci.tsv file directory if ccf data provided. Default NULL.}

\item{refBuild}{BSgenome.Hsapiens.UCSC reference. Default "hg19". Full genome sequences for Homo sapiens (Human) as provided by UCSC.}

\item{MafSummary}{Option for whether printing a MafSummary plot or not. Default TRUE.}

\item{outputDir}{Directory for ouput files. Default NULL.}

\item{gzOption}{Option for whether generating a .maf.gz compressed file. Default "". If a .maf.gz file is needed, the gzOption could be set as "gz"}
}
\value{
a classMaf object/class includes information of sample_info and 
mut.id and summary figure of it
}
\description{
Add sample information and ccf data to the original maf file 
to get a new maf. The new maf data adds three pieces of information:lesion, 
patient and time.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", 
                        package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", 
                               package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", 
                              package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", 
                           package = "Meskit")
## manually usage
maf <- readMaf(patientID="311252", mafFile=maf.File, 
               sampleInfoFile=sampleInfo.File, refBuild="hg19")
## if ccf data provided
maf <- readMaf(patientID="311252", mafFile=maf.File, 
               sampleInfoFile=sampleInfo.File, ccfClusterFile=pyCloneCluster, 
               ccfLociTsvFile=pyCloneLoci, refBuild="hg19")

}
