% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneFishplot.R
\name{cloneFishPlot}
\alias{cloneFishPlot}
\title{Get outputs from SCHISM and draw the fishplot}
\usage{
cloneFishPlot(maf = NULL, inferMethod = "clonevol",
  plotOption = "fishplot", schismCellularityFile = NULL,
  schismConsensusTree = NULL, genotypePosition = "space")
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{inferMethod}{choose a method to infer suboclonal relationship. Default "clonevol".}

\item{plotOption}{choose the function to generate different tyoe of fishplot. Default "fishplot".}

\item{schismCellularityFile}{specify the directory of cluster.cellularity document. Default NULL. If inferMethod == "SCHISM", the directory is essential and should be one of the outputs generated by  SCHISM.}

\item{schismConsensusTree}{specify the directory of GA.consensusTree document. Default NULL. However, if inferMethod == "SCHISM", the parameter should be one of the outputs generated by  SCHISM.}

\item{genotypePosition}{select corresponding clone positions for your timescape plot. Default "space". "stack" and "centre" are also available.}
}
\value{
the fishplot showing the evolution relationship predicted by SCHISM
}
\description{
infers subclonal relationship based on clonevol and SCHISM. 
generate metastatic evolvogram by fishplot and timescape.s
}
\examples{
\dontrun{
## clonevol method
cloneFishPlot(maf, inferMethod="clonevol", plotOption="fishplot")
cloneFishPlot(maf, inferMethod="clonevol", plotOption="timescape")
## SCHISM method
cloneFishPlot(inferMethod="SCHISM", plotOption="fishplot", schismCellularityFile=schismCellularityFile, schismConsensusTree=schismConsensusTree)
cloneFishPlot(inferMethod="SCHISM", plotOption="timescape", schismCellularityFile=schismCellularityFile, schismConsensusTree=schismConsensusTree)
}

}
