% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneFishplot.R
\name{cloneFishPlot}
\alias{cloneFishPlot}
\title{Get outputs from SCHISM and draw the fishplot}
\usage{
cloneFishPlot(inferMethod = "clonevol", plotOption = "fishplot",
  dirClusterTsvFile = NULL, dirClusterCellularity = NULL,
  dirGAconsensusTree = NULL, dirSampleInfo = NULL)
}
\arguments{
\item{inferMethod}{choose a method to infer suboclonal relationship. Default "clonevol".}

\item{plotOption}{choose the function to generate different tyoe of fishplot. Default "fishplot".}

\item{dirClusterTsvFile}{specify the directory of cluster.tsv file. Default NULL. If inferMethod == "clonevol", it is necessary to give the parameter a propper value.}

\item{dirClusterCellularity}{specify the directory of cluster.cellularity document. Default NULL. If inferMethod == "fishplot", the directory is essential and should be the output of SCHISM.}

\item{dirGAconsensusTree}{specify the directory of GA.consensusTree document. Default NULL. However, if inferMethod == "fishplot", the parameter should be the output of SCHISM.}

\item{dirSampleInfo}{pecify the directory of sample_info.txt. Default NULL.}
}
\value{
the fishplot showing the evolution relationship predicted by SCHISM
}
\description{

}
\examples{
\dontrun{
cloneFishPlot("SCHISM", "fishplot", dirClusterCellularity=dir.cluster.cellularity, dirGAconsensusTree=dir.GA.consensusTree)
cloneFishPlot("SCHISM", "timescape", dirClusterCellularity=dir.cluster.cellularity, dirGAconsensusTree=dir.GA.consensusTree, dirSampleInfo=NULL)
}

}
