% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{compareCCF}
\alias{compareCCF}
\title{compareCCF}
\usage{
compareCCF(
  maf,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.ccf = 0,
  pairByTumor = FALSE
)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.ccf}{The minimum value of CCF. Default: 0}

\item{pairByTumor}{Pair by tumor types in each patients,default is FALSE.}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
compareCCF(maf)
}
