% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfDensity.R
\name{compareCCF}
\alias{compareCCF}
\title{ccfPair}
\usage{
compareCCF(mafData, pairByType = FALSE, geneList = NULL, show.density = TRUE)
}
\arguments{
\item{show.density}{logical. If TRUE, the default, perform density estimation.}

\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.ccf}{the minimum value of CCF. Default: 0}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
ccfPair(maf, show.density = TRUE)
}
