% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{compareCCF}
\alias{compareCCF}
\title{compareCCF}
\usage{
compareCCF(
  maf,
  patient.id = NULL,
  pairByType = FALSE,
  geneList = NULL,
  min.ccf = 0,
  show.density = TRUE
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{pairByType}{pair by tumor types in each patients,default is FALSE.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{min.ccf}{the minimum value of CCF. Default: 0}

\item{show.density}{logical. If TRUE, the default, perform density estimation.}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
compareCCF(maf, show.density = TRUE)
}
