% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfDensity.R
\name{compareCCF}
\alias{compareCCF}
\title{ccfDensity}
\usage{
compareCCF(mafData, show.density = TRUE)
}
\arguments{
\item{show.density}{logical. If TRUE, the default, perform density estimation.}

\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
ccfDensity(maf, show.density = T)
}
