% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{compareCCF}
\alias{compareCCF}
\title{compareCCF}
\usage{
compareCCF(
  maf,
  patient.id = NULL,
  min.ccf = 0,
  pairByTumor = FALSE,
  use.tumorLabel = FALSE,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{min.ccf}{The minimum value of CCF. Default: 0}

\item{pairByTumor}{Pair by tumor types in each patients,default is FALSE.}

\item{use.tumorLabel}{Logical (Default: FALSE). Rename the 'Tumor_Sample_Barcode' with 'Tumor_Label'.}

\item{...}{Other options passed to \code{\link{subMaf}}}
}
\value{
a result list of CCF comparing between samples/tumor pairs
}
\description{
Compare the CCF between samples/tumor pairs
This function requires CCF for clustering
}
\examples{
maf.File <- system.file("extdata/", "HCC_LDC.maf", package = "MesKit")
clin.File <- system.file("extdata/", "HCC_LDC.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
compareCCF(maf)
}
