% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCCF.R
\name{compareCCF}
\alias{compareCCF}
\title{compareCCF}
\usage{
compareCCF(maf, patient.id = NULL, min.ccf = 0, pairByTumor = FALSE, ...)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{min.ccf}{The minimum value of CCF. Default: 0}

\item{pairByTumor}{Pair by tumor types in each patients,default is FALSE.}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\value{

}
\description{
Density plot of CCFs in paired samples of the patient
This function requires CCF for clustering
}
\examples{
compareCCF(maf)
}
