% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareJSI.R
\name{compareJSI}
\alias{compareJSI}
\title{compareJSI}
\usage{
compareJSI(
  maf,
  patient.id = NULL,
  pairByTumor = FALSE,
  min.ccf = 0,
  plot = TRUE,
  use.circle = TRUE,
  title = NULL,
  number.cex = 8,
  number.col = "#C77960",
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{pairByTumor}{Compare JSI between different tumors. Default: FALSE.}

\item{min.ccf}{The minimum value of CCF. Default: 0}

\item{plot}{Logical (Default:TRUE).}

\item{use.circle}{Logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix.}

\item{title}{Title of the plot, default is "Jaccard similarity".}

\item{number.cex}{The size of text shown in correlation plot. Default 8.}

\item{number.col}{The color of text shown in correlation plot. Default "#C77960".}

\item{...}{Other options passed to \code{\link{subMaf}}}
}
\value{
Correlation matrix and heatmap via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity index (JSI) is applied to distinguish monoclonal versus polyclonal seeding in metastases.
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
compareJSI(maf)
}
\references{
Hu, Z., Li, Z., Ma, Z. et al. Multi-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. Nat Genet (2020).
}
