% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareJSI.R
\name{compareJSI}
\alias{compareJSI}
\title{compareJSI}
\usage{
compareJSI(
  maf,
  patient.id = NULL,
  pairByTumor = FALSE,
  min.vaf = 0.02,
  plot = TRUE,
  use.circle = TRUE,
  title = NULL,
  number.cex = 8,
  number.col = "#C77960"
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.vaf}{specify the minimum VAF_adj, default is 0.08}

\item{plot}{logical (Default:TRUE). Whether to show the plot, default TRUE}

\item{use.circle}{logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix}

\item{title}{the title of the plot, default is "Jaccard similarity"}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSilent".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}
}
\value{
correlation matrix and heatmap via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity index (JSI) is applied to distinguish monoclonal versus polyclonal seeding in metastases
}
\examples{
compareJSI(maf)
}
\references{
Hu Z, Li Z, Ma Z, Curtis C: Pan-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. bioRxiv 2019.
}
