% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTree.R
\name{compareTree}
\alias{compareTree}
\title{compareTree}
\usage{
compareTree(
  tree1,
  tree2,
  plot = FALSE,
  min.ratio = 1/30,
  compare.linetype = "solid"
)
}
\arguments{
\item{tree1}{a phylogenetic tree}

\item{tree2}{a phylogenetic tree}

\item{plot}{a logical value. If TRUE, the two trees are plotted on the same device and their similarities are shown.}

\item{min.ratio}{a numeric value.If not NULL, min.length = max(tree$edge.length)*min.ratio.
Length of branches which are less than min.length will be adjusted to min.length.}

\item{compare.linetype}{line type of common branch.}
}
\value{
returns a vector containing the following tree distance methods by R package phangorn
Symmetric.difference  Robinson-Foulds distance
KF-branch distance  the branch score distance ( (Kuhner & Felsenstein 1994))
Path.difference  difference in the path length, counted as the number of branches 
Weighted.path.difference	 difference in the path length, counted using branches lengths
}
\description{
compares two phylogenetic trees and returns a detailed report of several distance methods
}
\examples{
tree1 <- getPhyloTree(maf, method = "NJ")
tree2 <- getPhyloTree(maf, method = "MP")
compareTree(tree1, tree2)
compareTree(tree1, tree2, plot = TRUE)
}
