% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTree.R
\name{compareTree}
\alias{compareTree}
\title{compareTree}
\usage{
compareTree(
  tree1,
  tree2,
  plot = FALSE,
  branchCol = "mutSig",
  show.heatmap = FALSE,
  min.ratio = 1/30,
  use.ccf = FALSE,
  show.bootstrap = FALSE,
  use.box = FALSE,
  common.lty = "solid"
)
}
\arguments{
\item{tree1}{a single class PhyloTree}

\item{tree2}{a single class PhyloTree}

\item{plot}{a logical value, default FALSE. If TRUE, 
the two trees will be plotted on the same device and their similarities will be shown.}

\item{min.ratio}{double (Default:1/30). If min.ratio is not NULL,
all edge length of a phylogenetic tree should be greater than
min.ratio*the longest edge length.
If not, the edge length will be reset as min.ratio*longest edge length.}

\item{common.lty}{line type of common branch.}
}
\value{
returns a vector containing the following tree distance methods by R package phangorn
Symmetric.difference  Robinson-Foulds distance
KF-branch distance  the branch score distance (Kuhner & Felsenstein 1994)
Path.difference  difference in the path length, counted as the number of branches 
Weighted.path.difference	 difference in the path length, counted using branches lengths
}
\description{
compares two phylogenetic trees and returns a detailed report of several distance methods
}
\examples{
tree1 <- getPhyloTree(maf, method = "NJ")[['patientID']]
tree2 <- getPhyloTree(maf, method = "MP")[['patientID']]
compareTree(tree1, tree2)
compareTree(tree1, tree2, plot = TRUE)
}
