% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTree.R
\name{compareTree}
\alias{compareTree}
\title{compareTree}
\usage{
compareTree(
  phyloTree1,
  phyloTree2,
  plot = FALSE,
  min.ratio = 1/20,
  show.bootstrap = FALSE,
  common.col = "red"
)
}
\arguments{
\item{phyloTree1}{A phyloTree object generated by \code{\link{getPhyloTree}} function.}

\item{phyloTree2}{A phyloTree object generated by \code{\link{getPhyloTree}} function.}

\item{plot}{FALSE(Default). If TRUE, two trees will be plotted on the same device and their similarities will be shown.}

\item{min.ratio}{Double (Default: 1/20). If min.ratio is not NULL,
all edge length which are smaller than min.ratio*the longest edge length will be reset as min.ratio*longest edge length.}

\item{show.bootstrap}{Logical. Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{common.col}{Color of common branches.}
}
\value{
A vector containing the following tree distance methods by R package phangorn
Symmetric.difference  Robinson-Foulds distance
KF-branch distance  the branch score distance (Kuhner & Felsenstein 1994)
Path.difference  difference in the path length, counted as the number of branches 
Weighted.path.difference	 difference in the path length, counted using branches lengths
}
\description{
Compares two phylogenetic trees and returns a detailed report of several distance methods
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")


phyloTree1 <- getPhyloTree(maf$HCC5647, method = "NJ")
phyloTree2 <- getPhyloTree(maf$HCC5647, method = "MP")
compareTree(phyloTree1, phyloTree2)
compareTree(phyloTree1, phyloTree2, plot = TRUE)
}
