% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTree.R
\name{compareTree}
\alias{compareTree}
\title{compareTree}
\usage{
compareTree(
  phyloTree1,
  phyloTree2,
  plot = FALSE,
  min.ratio = 1/20,
  show.bootstrap = FALSE,
  use.box = FALSE,
  common.col = "red"
)
}
\arguments{
\item{phyloTree1}{a phyloTree object generated by function getPhyloTree.}

\item{phyloTree2}{a phyloTree object generated by function getPhyloTree.}

\item{plot}{FALSE(Default). If TRUE, two trees will be plotted on the same device and their similarities will be shown.}

\item{min.ratio}{double (Default: 1/20). If min.ratio is not NULL,
all edge length which are smaller than min.ratio*the longest edge length will be reset as min.ratio*longest edge length.}

\item{common.col}{color of common branch.}
}
\value{
a vector containing the following tree distance methods by R package phangorn
Symmetric.difference  Robinson-Foulds distance
KF-branch distance  the branch score distance (Kuhner & Felsenstein 1994)
Path.difference  difference in the path length, counted as the number of branches 
Weighted.path.difference	 difference in the path length, counted using branches lengths
}
\description{
compares two phylogenetic trees and returns a detailed report of several distance methods
}
\examples{
phyloTree1 <- getPhyloTree(maf, method = "NJ", patient.id = 'patientID')
phyloTree2 <- getPhyloTree(maf, method = "MP", patient.id = 'patientID')
compareTree(phyloTree1, phyloTree2)
compareTree(phyloTree1, phyloTree2, plot = TRUE)
}
