% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutProfile.R
\name{genHeatmapPlotMatrix}
\alias{genHeatmapPlotMatrix}
\title{plotMutProfile}
\usage{
genHeatmapPlotMatrix(maf_data, topGenesCount = NULL)
}
\arguments{
\item{topGenesCount}{the number of genes print}

\item{maf}{the Maf object generated by readMaf
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{driverGenesFile}{driver gene list. Default NULL}
}
\description{
plotMutProfile
}
\examples{
plotMutProfile(maf, class = "SP", topGenesCount = 10)
}
