% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getBinaryMatrix}
\alias{getBinaryMatrix}
\alias{getBinaryMatrix,phyloTree-method}
\title{getBinaryMatrix}
\usage{
getBinaryMatrix(object)

\S4method{getBinaryMatrix}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Binary matrix of phyloTree
}
\description{
getBinaryMatrix
}
\examples{
getBinaryMatrix(phyloTree)
}
