% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getBootstrapValue}
\alias{getBootstrapValue}
\alias{getBootstrapValue,phyloTree-method}
\title{getBootstrapValue}
\usage{
getBootstrapValue(object)

\S4method{getBootstrapValue}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
bootstrap value of phyloTree
}
\description{
getBootstrapValue
}
\examples{
getBootstrapValue(phyloTree)
}
