% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMafData}
\alias{getMafData}
\alias{getMafData,Maf-method}
\title{getMafData}
\usage{
getMafData(object)

\S4method{getMafData}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Maf data
}
\description{
getMafData
}
\examples{
maf.File <- system.file("extdata/", "CRC_HZ.maf", package = "MesKit")
clin.File <- system.file("extdata/", "CRC_HZ.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "CRC_HZ.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
getMafData(maf$V402)
}
