% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMafPatient}
\alias{getMafPatient}
\alias{getMafPatient,Maf-method}
\title{getMafPatient}
\usage{
getMafPatient(object)

\S4method{getMafPatient}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Human reference genome versions of Maf
}
\description{
getMafPatient
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
getMafPatient(maf)
}
