% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutMatrix.R
\name{getMutMatrix}
\alias{getMutMatrix}
\title{getMutMatrix}
\usage{
getMutMatrix(mafData, use.ccf = FALSE)
}
\arguments{
\item{mafData}{the data information of the Maf object generated by readMaf function.}

\item{use.ccf}{whether to use the ccf data inthe mafData, default FALSE}
}
\value{
sorted binary or ccf mutation matrix
}
\description{
get the sorted binary or ccf mutation matrix
}
