% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJtree.R
\name{getNJtree}
\alias{getNJtree}
\title{Class NJtree}
\usage{
getNJtree(maf, use.indel = FALSE, ccf.mutation.id = c("Hugo_Symbol",
  "Chromosome", "Start_Position"), ccf.mutation.sep = ":")
}
\arguments{
\item{maf}{return from readMaf()}

\item{use.indel}{Seclet SNP in Variant type}

\item{ccf.mutation.id}{manually specify which columns could be joint by ccf.mutation.sep to get the same format of mutation id in ccfy}

\item{ccf.mutation.sep}{manually specify the separator character.Values on each line of the ccf.mutation.id are separated by this character. I}
}
\value{
NJtree object
}
\description{
Class NJtree
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "Meskit")
maf <- readMaf(patientID="311252", mafFile=maf.File, sampleInfoFile=sampleInfoFile, refBuild="hg19")
njtree <- getNJtree(maf)
}
