% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getNonSyn_vc}
\alias{getNonSyn_vc}
\alias{getNonSyn_vc,Maf-method}
\title{getNonSyn_vc}
\usage{
getNonSyn_vc(object)

\S4method{getNonSyn_vc}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
A list of Variant classifications which are considered as non-silent.
}
\description{
getNonSyn_vc
}
\examples{
getNonSyn_vc(maf)
}
