% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloTree.R
\name{getPhyloTree}
\alias{getPhyloTree}
\title{getPhyloTree}
\usage{
getPhyloTree(maf, method = "NJ", minVaf = 0.02, maxVaf = 1)
}
\arguments{
\item{method}{Approach to construct phylogenetic trees.Choose one of "NJ"(Neibor-Joining),"MP"(maximum parsimony),"ML"(maximum likelihood),""FASTME.ols" or "FASTME.bal".}

\item{minVaf}{the minimum value of vaf}

\item{maxVaf}{the maximum value of vaf}

\item{a}{Maf object generated by readMaf function}
}
\value{
an object of class PhyloTree
}
\description{
getPhyloTree
}
\examples{
maf.File <- system.file("extdata", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC6046.CCF.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
njtree <- getPhyloTree(maf)
}
