% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloTree.R
\name{getPhyloTree}
\alias{getPhyloTree}
\title{getPhyloTree}
\usage{
getPhyloTree(
  maf,
  patient.id = NULL,
  method = "NJ",
  min.vaf = 0.02,
  min.ccf = NULL,
  bootstrap.rep.num = 100,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE
)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{method}{Approach to construct phylogenetic trees.Choose one of "NJ"(Neibor-Joining),"MP"(maximum parsimony),"ML"(maximum likelihood),""FASTME.ols" or "FASTME.bal".}

\item{min.vaf}{The minimum value of vaf. Default 0.02.}

\item{min.ccf}{The minimum value of CCF. Default NULL.}

\item{bootstrap.rep.num}{Bootstrap iterations.. Default 100.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to include INDELs in analysis. Default: TRUE.}
}
\value{
PhyloTree or phyloTreeList object
}
\description{
getPhyloTree
}
\examples{
maf.File <- system.file("extdata", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC6046.CCF.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
}
