% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloTree.R
\name{getPhyloTree}
\alias{getPhyloTree}
\title{getPhyloTree}
\usage{
getPhyloTree(
  maf,
  method = "NJ",
  min.vaf = 0.02,
  min.ccf = NULL,
  bootstrap.rep.num = 100,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE
)
}
\arguments{
\item{method}{Approach to construct phylogenetic trees.Choose one of "NJ"(Neibor-Joining),"MP"(maximum parsimony),"ML"(maximum likelihood),""FASTME.ols" or "FASTME.bal".}

\item{min.vaf}{the minimum value of vaf}

\item{min.ccf}{the minimum value of CCF}

\item{bootstrap.rep.num}{bootstrap iterations.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{a}{Maf object generated by readMaf function}
}
\value{
a list of class PhyloTree
}
\description{
getPhyloTree
}
\examples{
maf.File <- system.file("extdata", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC6046.CCF.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
}
