% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPhyloTreeRef}
\alias{getPhyloTreeRef}
\alias{getPhyloTreeRef,phyloTree-method}
\title{getPhyloTreeRef}
\usage{
getPhyloTreeRef(object)

\S4method{getPhyloTreeRef}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Reference genome versions of phyloTree
}
\description{
getPhyloTreeRef
}
\examples{
maf.File <- system.file("extdata", "HCC_LDC.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile = maf.File, ccfFile = ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
getPhyloTreeRef(phyloTree$HCC5647)
}
