% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getSampleInfo}
\alias{getSampleInfo}
\alias{getSampleInfo,Maf-method}
\title{getSampleInfo}
\usage{
getSampleInfo(object)

\S4method{getSampleInfo}{Maf}(object)
}
\arguments{
\item{object}{An object of Maf}
}
\value{
Sample information
}
\description{
getSampleInfo
}
\examples{
maf.File <- system.file("extdata/", "CRC_HZ.maf", package = "MesKit")
clin.File <- system.file("extdata/", "CRC_HZ.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "CRC_HZ.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
getSampleInfo(maf$HCC5647)
}
