% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTree}
\alias{getTree}
\alias{getTree,phyloTree-method}
\title{getTree}
\usage{
getTree(object)

\S4method{getTree}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
Tree object of phyloTree
}
\description{
getTree
}
\examples{
maf.File <- system.file("extdata/", "HCC_LDC.maf", package = "MesKit")
clin.File <- system.file("extdata/", "HCC_LDC.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
getTree(phyloTree$HCC5647)
}
