% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTreeMethod}
\alias{getTreeMethod}
\alias{getTreeMethod,phyloTree-method}
\title{getTreeMethod}
\usage{
getTreeMethod(object)

\S4method{getTreeMethod}{phyloTree}(object)
}
\arguments{
\item{object}{An object of phyloTree}
}
\value{
tree construction method of phyloTree
}
\description{
getTreeMethod
}
\examples{
getTreeMethod(phyloTree)
}
