% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_input}
\alias{heatmap_input}
\title{plot the corresponding binary heatmap| ccf heatmaf of the NJtree}
\usage{
heatmap_input(mat, type)
}
\arguments{
\item{use.ccf}{logical. Whether to print the binary heatmap with ccf values. Defaults to FALSE, print the binary heatmap without ccf value}
}
\value{
binary mutation heatmap or mutation CCF heatmap
}
\description{
plot the corresponding binary heatmap| ccf heatmaf of the NJtree
}
\examples{
mut.heatmap(njtree)
mut.heatmap(njtree, use.ccf = TRUE)

}
