% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{JaccardIndex}
\alias{JaccardIndex}
\title{Visualize genetic heterogeneity between samples based on Jaccard Similarity coefficients of variants}
\usage{
JaccardIndex(maf, type = "lower")
}
\arguments{
\item{maf}{object generated by readMaf function}
}
\value{
the grapgical display of correlation matrix via Jaccard similarity coefficient method
'
}
\description{
The Jaccard similarity coefficient is defined as the ratio of shared variants to all 
(shared plus discordant) variants for two samples.
}
\examples{
JaccardIndex(maf)
}
