% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R, R/JaccardIndex_V1.R
\name{JSI_dist}
\alias{JSI_dist}
\title{calJSI}
\usage{
JSI_dist(df)

JSI_dist(df)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{plot}{logical (Default:TRUE). Whether to show the plot, default TRUE}

\item{use.circle}{logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix}

\item{title}{the title of the plot, default is "Jaccard similarity"}
}
\value{
correlation matrix and heatmap via Jaccard similarity coefficient method

correlation matrix and heatmap via Jaccard similarity coefficient method
}
\description{
The Jaccard similarity index (JSI) is applied to distinguish monoclonal versus polyclonal seeding in metastases

The Jaccard similarity index (JSI) is applied to distinguish monoclonal versus polyclonal seeding in metastases
}
\examples{
calJSI(maf)
calJSI(maf)
}
\references{
Hu Z, Li Z, Ma Z, Curtis C: Pan-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. bioRxiv 2019.

Hu Z, Li Z, Ma Z, Curtis C: Pan-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. bioRxiv 2019.
}
