% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classS4.R
\docType{class}
\name{MafList-class}
\alias{MafList-class}
\alias{MafList}
\title{MafList}
\description{
S4 class for storing a list of Maf objects.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a list of \code{\link{Maf}} objects.}
}}

\section{Constructor}{
\describe{
                 \item{\code{MafList (...)}}{combine multiple Maf 
                 objects supplied in ... into a MafList object.}
               }
}

