% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{mathScore}
\alias{mathScore}
\title{mathScore}
\usage{
mathScore(
  maf,
  patient.id = NULL,
  min.vaf = 0.02,
  withinType = FALSE,
  use.adjVAF = FALSE
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.vaf}{the minimum VAF for filtering variants. Default: 0.02}

\item{withinType}{calculate AUC within types in each patients, default is FALSE.}
}
\value{
a data.table of MATH scores
}
\description{
calculates MATH score of each tumor sample or based on Mutant-Allele Tumor Heterogeneity (MATH) approach.
}
\examples{
mathScore(maf, min.vaf = 0.02)
}
\references{
Mroz, Edmund A. et al. Intra-Tumor Genetic Heterogeneity and Mortality in Head and Neck Cancer: Analysis of Data from The Cancer Genome Atlas. Ed. Andrew H. Beck. PLoS Medicine 12.2 (2015): e1001786.
}
