% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{mathScore}
\alias{mathScore}
\title{A Multifunctional MATH Calculator}
\usage{
mathScore(maf, tsb = c("All"), minvaf = 0.02, maxvaf = 1)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{tsb}{specify excute ITH asessment for single or all samples from the same patient(Tumor_Sample_Barcodes)}

\item{minvaf}{filter low frequency variants caused by sequencing error. Default 0. (on the scale of 0 to 1)}

\item{maxvaf}{filter high frequency variants due to copy number alterations or impure tumor. Default 1. (on the scale of 0 to 1)}
}
\value{
MATH score list for all/selected samples or a MATH score value.
}
\description{
Calculate MATH score and present results in different ways 
determined by parameters. VAF can be on the scale 0-1 and you could filter 
VAF with minvaf and maxvaf.
}
\examples{
mathScore(maf, tsb=c("All"), minvaf=0.02, maxvaf=1)
## generate MATH score for specific tsbs
mathScore(maf, tsb=c("SRR3670035"))

}
