% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathScore.R
\name{mathScore}
\alias{mathScore}
\title{mathScore}
\usage{
mathScore(maf, tsb = NULL, minvaf = 0.02, maxvaf = 1)
}
\arguments{
\item{maf}{a Maf object generated by readMaf.}

\item{tsb}{Specify Tumor_Sample_Barcodes for single or all samples from maf of the patient. Default: NULL, which means that output all samples' MATH Score.}

\item{minvaf}{the minimum variant allele frequencies (VAFs) for filtering variants. Default: 0. Option: on the scale of 0 to 1.}

\item{maxvaf}{the maximum variant allele frequencies (VAFs) for filtering variants. Default: 0. Option: on the scale of 0 to 1.}
}
\value{
A list containing MATH scores of all/selected samples.
}
\description{
Calculate MATH score and VAF value could be filtered with minvaf and maxvaf.
}
\examples{
mathScore(maf, tsb=NULL, minvaf=0.02, maxvaf=1)
mathScore(maf, tsb=c("SRR3670035"))

}
