% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutHeatmap.R
\name{mutHeatmap}
\alias{mutHeatmap}
\title{mutHeatmap}
\usage{
mutHeatmap(
  maf,
  use.ccf = FALSE,
  show.class.label = TRUE,
  geneList = NULL,
  plot.geneList = FALSE,
  show.gene = FALSE,
  show.geneList = TRUE,
  patient.id = NULL,
  mut.threshold = 150
)
}
\arguments{
\item{maf}{Maf object generated by readMaf function}

\item{use.ccf}{logical. If FALSE (default), print a binary heatmap of mutations. Otherwise, print a cancer cell frequency (CCF) heatmap.}

\item{show.class.label}{logical. If TRUE, show labels for mutations classes on the heatmap.Default is TRUE.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{plot.geneList}{if TRUE, plot heatmap with genes on geneList when geneList is not NULL.Default FALSE.}

\item{show.gene}{show the name of genes next to the heatmap.Default FALSE.}

\item{show.geneList}{show the names of gene on the geneList.Default FALSE.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included.}

\item{mut.threshold}{show.gene and show.geneList will be FALSE when patient have more mutations than threshold.Default is 150.}
}
\value{
heatmap of somatic mutations
}
\description{
plot the corresponding binary heatmap or ccf heatmaf of the phyloTree
}
\examples{
mutHeatmap(phyloTree)
mutHeatmap(phyloTree, use.ccf = TRUE)

}
