% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutHeatmap.R
\name{mutHeatmap}
\alias{mutHeatmap}
\title{mutHeatmap}
\usage{
mutHeatmap(
  maf,
  min.vaf = 0.02,
  use.ccf = FALSE,
  patient.id = NULL,
  geneList = NULL,
  plot.geneList = FALSE,
  show.gene = FALSE,
  show.geneList = TRUE,
  mut.threshold = 50,
  class.text.size = 4,
  sample.text.size = 9,
  legend.title.size = 10,
  gene.text.size = 9
)
}
\arguments{
\item{maf}{Maf object generated by readMaf function}

\item{min.vaf}{the minimum value of VAF. Default: 0. Option: on the scale of 0 to 1.}

\item{use.ccf}{logical. If FALSE (default), print a binary heatmap of mutations. Otherwise, print a cancer cell frequency (CCF) heatmap.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{plot.geneList}{if TRUE, plot heatmap with genes on geneList when geneList is not NULL.Default FALSE.}

\item{show.gene}{show the name of genes next to the heatmap.Default FALSE.}

\item{show.geneList}{show the names of gene on the geneList.Default FALSE.}

\item{mut.threshold}{show.gene and show. geneList will be FALSE when patient have more mutations than threshold.Default is 150.}
}
\value{
heatmap of somatic mutations
}
\description{
plot binary or CCF heatmap of somatic mutations.
}
\examples{
mutHeatmap(phyloTree)
mutHeatmap(phyloTree, use.ccf = TRUE)
}
