% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutHeatmap.R
\name{mutHeatmap}
\alias{mutHeatmap}
\title{mutHeatmap}
\usage{
mutHeatmap(phyloTree, use.ccf = FALSE)
}
\arguments{
\item{use.ccf}{logical. If FALSE (default), print a binary heatmap of mutations. Otherwise, print a cancer cell frequency (CCF) heatmap.}

\item{Maf}{an \code{\link{maf}} object generated by \code{\link{read.maf}}}
}
\value{
heatmap of somatic mutations
}
\description{
plot the corresponding binary heatmap or ccf heatmaf of the phyloTree
}
\examples{
mutHeatmap(phyloTree)
mutHeatmap(phyloTree, use.ccf = TRUE)

}
