% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutOncoTSG.R
\name{mutOncoTSG}
\alias{mutOncoTSG}
\title{mutOncoTSG}
\usage{
mutOncoTSG(maf, oncogeneListFile, tsgListFile, show.percentage = FALSE)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{oncogeneListFile}{oncogenes list.}

\item{tsgListFile}{TSGs list.}

\item{show.percentage}{logical parameter that determine whether the 
percentage of different intersections will print in the stack plot or not}
}
\description{
With the gene list passed by oncogeneList and tsgList, 
mutOncoTSG can calculate the proportion of shared/partial-shared/private variants
that occurs in oncogenes and tumor suppressors genes(TSGs)
1. Private: variants specific for each samples; 2. Parital-shared: variants shared 
by multiple samples but not existing in all samples; 
3. Shared: variants shared by all samples.
}
\examples{
maf.File <- system.file("extdata/maf", "HCC6046.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata", "HCC6046.sampleInfo.txt", package = "MesKit")
maf <- readMaf(mafFile = maf.File, sampleInfo = sampleInfo.File)
oncogeneListFile <- system.file("extdata/", "oncogene.list.txt", package = "MesKit")
tsgListFile <- system.file("extdata/", "TSG.list.txt", package = "MesKit")
mutOncoTSG(maf, oncogeneListFile, tsgListFile, show.percentage=TRUE)

}
