% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutSharedPrivate.R
\name{mutSharedPrivate}
\alias{mutSharedPrivate}
\title{Use R code to find the intersect mutations and their types in several samples of one patient}
\usage{
mutSharedPrivate(maf, show.num = FALSE)
}
\arguments{
\item{maf}{Maf object return from read.Maf()}

\item{show.num}{a logic parameter to determine whether to show the number of each mutations in the stack plot}

\item{patientID}{the ID of a patient whose cancer information is going to be analyzed}
}
\value{
Stack and point-line plot
}
\description{
Use R code to find the intersect mutations and their types in several samples of one patient
}
\examples{
maf <- read.Maf('patient', dat.dir = './data/multi_lesion' )
mutSharedPrivate(maf, 311252)
}
