% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutSharedPrivate.R
\name{mutSharedPrivate}
\alias{mutSharedPrivate}
\title{Use R code to find the intersect mutations and their types in several samples of one patient}
\usage{
mutSharedPrivate(maf, showNumber = FALSE)
}
\arguments{
\item{maf}{Maf object return from read.Maf()}

\item{showNumber}{a logic parameter to determine whether to show the number of each mutations in the stack plot}
}
\description{
Use R code to find the intersect mutations and their types in several samples of one patient
}
\examples{
maf.File <- system.file("extdata/maf", "HCC6046.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata", "HCC6046.sampleInfo.txt", package = "Meskit")
ccf.cluster.File <- system.file("extdata/ccf", "HCC6046.cluster.tsv", package = "Meskit")
ccf.loci.File <- system.file("extdata/ccf", "HCC6046.loci.tsv", package = "Meskit")
maf <- readMaf(mafFile = maf.File, sampleInfo = sampleInfo.File, 
                ccfClusterTsvFile = ccf.cluster.File, 
                ccfLociTsvFile = ccf.loci.File,
                refBuild = "hg19")
mutSharedPrivate(maf)
}
