% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R, R/mutTrunkBranch.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(treeMSOutput, conf.level = 0.95)

mutTrunkBranch(treeMSOutput, conf.level = 0.95)
}
\arguments{
\item{treeMSOutput}{The output of function treeMutationalSig.}

\item{conf.level}{The confidence level to vertify whether the kind of mutation is significant. Default: 0.95. Option: on the scale of 0 to 1.}

\item{treeMSOutput}{the output of treeMutationalSig function.}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
Box plots based on mutational categories

Box plots based on mutational categories
}
\description{
Provide a summary data frame for mutations in trunk or branches in phylotree.

Summarize and conduct paired Wilcoxon test of mutations of trunk/branches in a phylogenetic tree.
}
\examples{
mutTrunkBranch(treeMSOutput, conf.level = 0.95)

mutTrunkBranch(treeMSOutput, conf.level = 0.95)
}
