% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTrunkBranch.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(
  phyloTree,
  geneList = NULL,
  signaturesRef = "cosmic",
  plot = TRUE
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by phyloTree function.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic". Option: "nature2013".}

\item{plot}{output a list of diagrams that visualize mutational signature of trunk/branches in a phylogenetic tree.Default FALSE.}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
a list of box plots based on mutational categories
}
\description{
Summarize and conduct paired Wilcoxon test of mutations of trunk/branches in a phylogenetic tree.
}
\examples{
mutTrunkBranch(phyloTree)
mutTrunkBranch(phyloTree,plot = T)

}
