% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTrunkBranch.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(
  phyloTree_list,
  CT = FALSE,
  geneList = NULL,
  pvalue = 0.05,
  plot = TRUE,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree_list}{a list of phyloTree generated by phyloTree function.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{plot}{output a list of diagrams that visualize mutational signature of trunk/branches in a phylogenetic tree.Default FALSE.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
a list of box plots based on mutational categories
}
\description{
Summarize and conduct paired Wilcoxon test of mutations of trunk/branches in a phylogenetic tree.
}
\examples{
mutTrunkBranch(phyloTree_list)
mutTrunkBranch(phyloTree_list,plot = T)

}
