% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(treeMSOutput, conf.level = 0.95)
}
\arguments{
\item{treeMSOutput}{The output of function treeMutationalSig.}

\item{conf.level}{The confidence level to vertify whether the kind of mutation is significant. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
Box plots based on mutational categories
}
\description{
Draw box plots based on mutational categories
}
\examples{
mutTrunkBranch(treeMSOutput, conf.level = 0.95)

}
