% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTrunkBranch.R
\name{mutTrunkBranch}
\alias{mutTrunkBranch}
\title{mutTrunkBranch}
\usage{
mutTrunkBranch(tree.mutSig, conf.level = 0.95)
}
\arguments{
\item{tree.mutSig}{the output of treeMutSig function.}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
Box plots based on mutational categories
}
\description{
Summarize and conduct paired Wilcoxon test of mutations of trunk/branches in a phylogenetic tree.
}
\examples{
mutTrunkBranch(tree.mutSig, conf.level = 0.95)
}
