% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutProfile.R
\name{mutType}
\alias{mutType}
\title{plotMutProfile}
\usage{
mutType(unique_sample_count, Status, total_count)
}
\arguments{
\item{maf}{the Maf object generated by readMaf}

\item{driverGenesFile}{driver gene list. Default NULL}

\item{topGenesCount}{the number of genes print}
}
\description{
plotMutProfile
}
\examples{
plotMutProfile(maf, topGenesCount = 10)
}
