% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMut.R
\name{mutType_SP}
\alias{mutType_SP}
\title{classifyMut}
\usage{
mutType_SP(unique_sample_count, total_count)
}
\arguments{
\item{maf}{the Maf object generated by readMaf
Classify SSNVs/Indels into Shared/P-shared/Private, Clonal/Subclonl
or Shared-Clonal/P-shared-Clonal/Private-Clonal/Shared-Subclonal/P-shared-SubClonal/Private-SubClonal}

\item{driverGenesFile}{driver gene list. Default NULL}
}
\description{
classifyMut
}
\examples{
classifyMut(maf, class = "SP")
}
