% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNeiDist.R
\name{nei_dist}
\alias{nei_dist}
\title{calFst}
\usage{
nei_dist(df, withinType = FALSE)
}
\arguments{
\item{maf}{an Maf object generated by readMaf function}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{min.vaf}{specify the minimum VAF, default is 0.08}

\item{max.vaf}{specify the minimum VAF, default is 1}

\item{plot}{logical (Default:TRUE). Whether to show the plot, default TRUE}

\item{use.circle}{logical (Default:TRUE). Whether to use "circle" as visualization method of correlation matrix
default TURE}

\item{title}{the title of the plot, default is "Nei's distance"}
}
\value{
Nei's genetic distance matrix and heatmap of sample-pairs from the same patient
}
\description{
Nei's distance of CCF for each sample-pair
}
\examples{
calNeiDist(maf)
}
\references{
Lee JK, Wang J, Sa JK, et al. Spatiotemporal genomic architecture informs precision oncology in glioblastoma. 
Nat Genet. 2017;49(4):594–599. doi:10.1038/ng.3806
}
