% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJtree.R, R/NJtree2.R
\name{NJtree}
\alias{NJtree}
\title{Class NJtree}
\usage{
NJtree(maf, use.indel = FALSE, mut.signature = TRUE,
  sig.min.mut.number = 50, ccf.mutation.id, ccf.mutation.sep)

NJtree(maf, use.indel = FALSE, mut.signature = TRUE,
  sig.min.mut.number = 50, ccf.mutation.id, ccf.mutation.sep)
}
\arguments{
\item{maf}{return from readMaf()}

\item{use.indel}{Seclet SNP in Variant type}

\item{sig.min.mut.number}{minimum mutation number in each branch}

\item{ccf.mutation.id}{determine the id of mutation}

\item{ccf.mutation.sep}{sep of id}

\item{maf}{return from readMaf()}

\item{use.indel}{Seclet SNP in Variant type}

\item{sig.min.mut.number}{minimum mutation number in each branch}

\item{ccf.mutation.id}{determine the id of mutation}

\item{ccf.mutation.sep}{sep of id}
}
\value{
NJtree object

NJtree object
}
\description{
Class NJtree

Class NJtree
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "Meskit")
maf <- readMaf(patientID = "311252", mafFile = maf.File, sampleInfoFile = sampleInfo.File, refBuild = "hg19")
njtree <- NJtree(maf)
getNJtreenj(njtree)
getNJtreemut_sort(njtree)
getNJtreeSignature(njtree)
maf <- readMaf(patientID = "311252",mafDir = './inst/exdata/multi_lesion/maf/311252.maf', refBuild = "hg19")
njtree <- read.NJtree(maf)
getNJtreenj(njtree)
getNJtreemut_sort(njtree)
getNJtreeSignature(njtree)
}
