% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNA.R
\name{plotCNA}
\alias{plotCNA}
\title{plotCNA}
\usage{
plotCNA(seg, sample.as.col = FALSE, show.GISTIC.gene = FALSE)
}
\arguments{
\item{seg}{object generated by readSegment function.}

\item{sample.as.col}{Whether to use samples as each column of the heat map.Default FALSE.}

\item{show.GISTIC.gene}{Whether GISTIC gene in seg.Default FALSE.}
}
\description{
plotCNA
}
\examples{
segCN.file <- system.file("extdata", "HCC6046.seg.txt", package = "MesKit")
# Load gtf.dat
load(system.file("extdata", "Homo_sapiens.GRCh37.75.RData", package = "MesKit"))
gisticAmpGenesFile <- system.file("extdata", "LIHC_amp_genes.conf_99.txt", package = "MesKit")
gisticDelGenesFile <- system.file("extdata", "LIHC_del_genes.conf_99.txt", package = "MesKit")
seg <- readSegment(segCN.file = segCN.file, gtf.dat = gtf.dat, gisticAmpGenesFile = gisticAmpGenesFile, gisticDelGenesFile = gisticDelGenesFile)
plotCNA(seg)

}
