% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNA.R
\name{plotCNA}
\alias{plotCNA}
\title{plotCNA}
\usage{
plotCNA(
  seg,
  patient.id = NULL,
  refBuild = "hg19",
  chrSilent = NULL,
  sample.fontsize = 11,
  chrom.fontsize = 3,
  legend.text.fontsize = 9,
  legend.title.fontsize = 11,
  sample.bar.height = 0.5,
  chrom.bar.height = 0.5,
  show.GISTIC.gene = FALSE
)
}
\arguments{
\item{seg}{object generated by readSegment function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included.}

\item{refBuild}{human reference genome versions of hg18, hg19 or hg38 by UCSC. Default: "hg19".}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, chrX, chrY. Default NULL.}

\item{show.GISTIC.gene}{Whether GISTIC gene in seg.Default FALSE.}
}
\description{
plotCNA
}
\examples{
segCN.file <- system.file("extdata", "HCC6046.seg.txt", package = "MesKit")
seg <- readSegment(segCN.file = segCN.file)
plotCNA(seg)

}
