% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNA.R
\name{plotCNA}
\alias{plotCNA}
\title{plotCNA}
\usage{
plotCNA(
  seg,
  patient.id = NULL,
  sampleOrder = NULL,
  chrSilent = NULL,
  refBuild = "hg19",
  sample.text.size = 11,
  chrom.text.size = 3,
  legend.text.size = 9,
  legend.title.size = 11,
  sample.bar.height = 0.5,
  chrom.bar.height = 0.5,
  showRownames = TRUE,
  removeEmptyChr = TRUE,
  showCytoband = FALSE,
  showGene = FALSE,
  use.tumorLabel = FALSE
)
}
\arguments{
\item{seg}{Object generated by \code{\link{readSegment}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{sampleOrder}{A named list which contains the sample order used in plotting the final profile. Default: NULL}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, 1, 2, 3. Default NULL.}

\item{refBuild}{Human reference genome versions of hg18, hg19 or hg38 by UCSC. Default: "hg19".}

\item{sample.text.size}{Size of sample name.Default 11.}

\item{chrom.text.size}{Size of chromosome text.Default 3.}

\item{legend.text.size}{Size of legend text. Default 9.}

\item{legend.title.size}{Size of legend title.Default 11.}

\item{sample.bar.height}{Bar height of each sample .Default 0.5.}

\item{chrom.bar.height}{Bar height of each chromosome .Default 0.5.}

\item{showRownames}{Show sample names of rows.Default is TRUE.}

\item{removeEmptyChr}{Remove empty chromosomes that do not exist in all samples.Default is TRUE.}

\item{showCytoband}{Show the information about cytoband on the plot.Default is FALSE}

\item{showGene}{Show the information about Gene on the plot if parameter 'txdb' is not null in function readSegment.Default is FALSE}

\item{use.tumorLabel}{Logical (Default: FALSE). Rename the 'Tumor_Sample_Barcode' with 'Tumor_Label'.}
}
\value{
a heatmap plot of CNA profile
}
\description{
plotCNA
}
\examples{
segFile <- system.file("extdata", "HCC_LDC.seg.txt", package = "MesKit")
seg <- readSegment(segFile = segFile)
plotCNA(seg)

## showCytoband
segFile <- system.file("extdata", "HCC_LDC.seg.txt", package = "MesKit")
gisticAmpGenesFile <- system.file("extdata", "LIHC_amp_genes.conf_99.txt", package = "MesKit")
gisticDelGenesFile <- system.file("extdata", "LIHC_del_genes.conf_99.txt", package = "MesKit")
gisticAllLesionsFile <- system.file("extdata", "LIHC_all_lesions.conf_99.txt", package = "MesKit")
seg <- readSegment(segFile = segFile,
                   gisticAmpGenesFile = gisticAmpGenesFile,
                    gisticDelGenesFile = gisticDelGenesFile,
                   gisticAllLesionsFile = gisticAllLesionsFile)
plotCNA(seg, showCytoband = TRUE)

## showGene
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(org.Hs.eg.db)
library(AnnotationDbi)
txdb = TxDb.Hsapiens.UCSC.hg19.knownGene

segFile <- system.file("extdata", "HCC_LDC.seg.txt", package = "MesKit")
gisticAmpGenesFile <- system.file("extdata", "LIHC_amp_genes.conf_99.txt", package = "MesKit")
gisticDelGenesFile <- system.file("extdata", "LIHC_del_genes.conf_99.txt", package = "MesKit")
gisticAllLesionsFile <- system.file("extdata", "LIHC_all_lesions.conf_99.txt", package = "MesKit")

seg <- readSegment(segFile = segFile,
                   gisticAmpGenesFile = gisticAmpGenesFile,
                    gisticDelGenesFile = gisticDelGenesFile, 
                   gisticAllLesionsFile = gisticAllLesionsFile,
                   txdb = txdb)

plotCNA(seg, showGene = TRUE)
                   

}
