% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNA.R
\name{plotCNA}
\alias{plotCNA}
\title{plotCNA}
\usage{
plotCNA(
  seg,
  patient.id = NULL,
  sampleOrder = NULL,
  chrSilent = NULL,
  refBuild = "hg19",
  sample.text.size = 11,
  chrom.text.size = 3,
  legend.text.size = 9,
  legend.title.size = 11,
  sample.bar.height = 0.5,
  chrom.bar.height = 0.5,
  showRownames = TRUE
)
}
\arguments{
\item{seg}{Object generated by \code{\link{readSegment}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{sampleOrder}{A named list which contains the sample order used in plotting the final profile. Default: NULL}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, 1, 2, 3. Default NULL.}

\item{refBuild}{Human reference genome versions of hg18, hg19 or hg38 by UCSC. Default: "hg19".}

\item{sample.text.size}{Size of sample name.Default 11.}

\item{chrom.text.size}{Size of chromosome text.Default 3.}

\item{legend.text.size}{Size of legend text. Default 9.}

\item{legend.title.size}{Size of legend title.Default 11.}

\item{sample.bar.height}{Bar height of each sample .Default 0.5.}

\item{chrom.bar.height}{Bar height of each chromosome .Default 0.5.}

\item{showRownames}{show sample names.Default TRUE.}
}
\value{
a heatmap plot of CNA profile
}
\description{
plotCNA
}
\examples{
segFile <- system.file("extdata", "HCC_LDC.seg.txt", package = "MesKit")
seg <- readSegment(segFile = segFile)
plotCNA(seg)

}
