% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNA.R
\name{plotCNA}
\alias{plotCNA}
\title{plotCNA}
\usage{
plotCNA(seg, refBuild = "hg19", show.GISTIC.gene = FALSE)
}
\arguments{
\item{seg}{object generated by readSegment function.}

\item{refBuild}{human reference genome versions of hg18, hg19 or hg38 by UCSC. Default "hg19".}

\item{show.GISTIC.gene}{Whether GISTIC gene in seg.Default FALSE.}
}
\description{
plotCNA
}
\examples{
seg <- readSegment(segCN.file = segCN.file)
plotCNA(seg)

}
