% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfile}
\alias{plotMutSigProfile}
\title{plotMutSigProfile}
\usage{
plotMutSigProfile(sig_input, patient.id = NULL, mode = NULL)
}
\arguments{
\item{patient.id}{patient.id select the specific patients. Default: NULL, all patients are included.}

\item{mode}{type of mutation spectrum.Default: NULL. Options:'Original','Reconstructed' or 'Difference'}

\item{sig_out}{result generated by function fitSignatures or triMatrix.}
}
\value{
mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
## input from fitSignatures
tri_matrix <- triMatrix(phyloTree)
fit_out <- fitSignatrues(tri_matrix)
plotMutSigProfile(fit_out)
## input from treeMatrix
plotMutSigProfile(tri_matrix)

}
