% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfile}
\alias{plotMutSigProfile}
\title{plotMutSigProfile}
\usage{
plotMutSigProfile(
  sig_input,
  patient.id = NULL,
  mode = NULL,
  use.Tumor_Label = FALSE
)
}
\arguments{
\item{sig_input}{Result generated by function \code{\link{fitSignatures}} or \code{\link{triMatrix}}.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{mode}{Type of mutation spectrum.Default: NULL. Options:'Original','Reconstructed' or 'Difference'}

\item{use.Tumor_Label}{Let Tumor_Sample_Barcode be Tumor_Label if Tumor Label is provided in clinical data.Default FALSE.}
}
\value{
Mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
## input from fitSignatures
maf.File <- system.file("extdata/", "HCC_LDC.maf", package = "MesKit")
clin.File <- system.file("extdata/", "HCC_LDC.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf, patient.id = 'HCC8257')

## Load a reference genome.
library(BSgenome.Hsapiens.UCSC.hg19)

tri_matrix <- triMatrix(phyloTree)
fit_out <- fitSignatures(tri_matrix)
plotMutSigProfile(fit_out)
## input from treeMatrix
plotMutSigProfile(tri_matrix)

}
