% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfile}
\alias{plotMutSigProfile}
\title{plotMutSigProfile}
\usage{
plotMutSigProfile(treeMutSig_output, patient.id = NULL, mode = NULL)
}
\arguments{
\item{treeMutSig_output}{object generated by treeMutSig function.}

\item{patient.id}{patient.id select the specific patients. Default: NULL, all patients are included.}

\item{mode}{type of mutation spectrum.Default: NULL. Options:'Original','Reconstructed' or 'Difference'}
}
\value{
a list of mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
treeMutSig_output <- treeMutSig(phyloTree)
plotMutSigProfile(treeMutSig_output)
}
