% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfile}
\alias{plotMutSigProfile}
\title{plotMutSigProfile}
\usage{
plotMutSigProfile(sig_input, patient.id = NULL, mode = NULL)
}
\arguments{
\item{sig_input}{Result generated by function \code{\link{fitSignatures}} or \code{\link{triMatrix}}.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{mode}{Type of mutation spectrum.Default: NULL. Options:'Original','Reconstructed' or 'Difference'}
}
\value{
Mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
## input from fitSignatures
maf.File <- system.file("extdata", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC6046.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
phyloTree <- getPhyloTree(maf)
tri_matrix <- triMatrix(phyloTree)
fit_out <- fitSignatrues(tri_matrix)
plotMutSigProfile(fit_out)
## input from treeMatrix
plotMutSigProfile(tri_matrix)

}
