% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutSigProfile.R
\name{plotMutSigProfiler}
\alias{plotMutSigProfiler}
\title{plotMutSigProfile}
\usage{
plotMutSigProfiler(mutSig.spectrum, patient.id = NULL, mode = NULL)
}
\arguments{
\item{mutSig.spectrum}{a list of probability of 96 mutation types generated by treeMutSig function.}

\item{patient.id}{patient.id select the specific patients. Default: NULL, all patients are included.}
}
\value{
a list of mutation signature profile of patients
}
\description{
plotMutSigProfile
}
\examples{
tree.mutSig <- treeMutSig(phyloTree)
plotMutSigProfiler(tree.mutSig$mutSig.spectrum)
}
