% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree = NULL,
  show.mutSig = TRUE,
  show.heatmap = TRUE,
  heatmap.type = "binary",
  use.box = TRUE
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getphyloTree function.}

\item{show.mutSig}{logical. Whether to show mutational signature on tree. Default is TRUE.}

\item{show.heatmap}{logical. Whether to show heatmap of somatic mutations. Default is TRUE.}

\item{heatmap.type}{character. "binary" (default) for printing a binary heatmap of mutations; or "CCF" for printing a cancer cell frequency (CCF) heatmap. This parameter is only useful when show.mutSig = TRUE.}

\item{use.box}{logical.Whether to add box arround bootstrap value on tree. Default is TRUE.}
}
\value{
Prints plot.
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree)
## Use ccf 
plotPhyloTree(phyloTree, heatmap.type = 'CCF')
}
