% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(njtree = NULL, show.mutSig = TRUE, show.heatmap = TRUE,
  heatmap.type = "binary")
}
\arguments{
\item{njtree}{a njtree object generated by NJtree function.}

\item{show.mutSig}{logical. Whether to show mutational signature on tree. Default is TRUE.}

\item{show.heatmap}{logical. Whether to show heatmap of somatic mutations. Default is TRUE.}

\item{heatmap.type}{character. "binary" (default) for printing a binary heatmap of mutations; or "CCF" for printing a cancer cell frequency (CCF) heatmap. This parameter is only useful when show.mutSig = TRUE.}
}
\value{
Prints plot.
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(njtree)
## Use ccf 
plotPhyloTree(njtree, heatmap.type = 'CCF')
}
