% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{A phylogenetic tree painter}
\usage{
plotPhyloTree(
  njtree = NULL,
  showMutSig = TRUE,
  sigName = "default",
  showHeatmap = TRUE,
  heatmapType = "binary"
)
}
\arguments{
\item{njtree}{a njtree object generated by NJtree function.}

\item{showMutSig}{logical. Whether to show mutational signature on tree. Default is TRUE.}

\item{sigName}{Choose "alias" when you want showing alias in branch.}

\item{showHeatmap}{logical. Whether to show heatmap of somatic mutations. Default is TRUE.}

\item{heatmapType}{character. "binary" (default) for printing a binary heatmap of mutations; or "CCF" for printing a cancer cell frequency (CCF) heatmap. This parameter is only useful when show.mutSig = TRUE.}
}
\value{
Prints plot.
}
\description{
A phylogenetic tree painter
}
\examples{
plotPhyloTree(njtree)
## Use ccf 
plotPhyloTree(njtree, heatmap.type = 'CCF')
}
