% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree_list = NULL,
  patient.id = NULL,
  branchCol = "mutSig",
  show.bootstrap = TRUE,
  use.box = TRUE,
  common.lty = "solid",
  min.ratio = 1/30,
  min.mut.count = 15,
  signaturesRef = "cosmic_v2",
  compare = FALSE,
  treeData = NULL
)
}
\arguments{
\item{phyloTree_list}{a list of phyloTree object generated by getphyloTree function.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}

\item{branchCol}{character."mutSig" for coloring branches by mutation signature; "mutType" for coloring branches by branch tumor type; NULL for coloring automatically.Default is 'mutSig'.}

\item{show.bootstrap}{logical.Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{use.box}{logical.Whether to add box around bootstrap value on tree. Default is TRUE.}

\item{common.lty}{common.lty line type of common branch which is used for compareTree.

are mapped along the trees as indicated}

\item{min.ratio}{double (Default:1/30). If min.ratio is not NULL,
all edge length of a phylogenetic tree should be greater than
min.ratio*the longest edge length.
If not, the edge length will be reset as min.ratio*longest edge length.}

\item{min.mut.count}{the threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{compare}{parameters for compareTree.}

\item{treeData}{parameters for compareTree.}
}
\value{
return a list of phylotree graph .
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree_list)
}
