% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree = NULL,
  branchCol = "mutSig",
  show.heatmap = TRUE,
  heatmap.type = "binary",
  show.bootstrap = TRUE,
  use.box = TRUE,
  min.ratio = 1/30,
  show.class.label = FALSE,
  geneList = NULL,
  plot.geneList = FALSE,
  show.gene = FALSE,
  show.geneList = TRUE,
  mut.threshold = 150,
  min.mut.count = 15,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getphyloTree function.}

\item{branchCol}{character."mutSig" for colouring branches by mutation signature;"mutType" for colouring branches by branch tumor type;NULL for colouring automatically.Default is 'mutSig'.}

\item{show.heatmap}{logical. Whether to show heatmap of somatic mutations. Default is TRUE.}

\item{heatmap.type}{character. "binary" (default) for printing a binary heatmap of mutations; or "CCF" for printing a cancer cell frequency (CCF) heatmap. This parameter is only useful when branchCol = "mutSig".}

\item{show.bootstrap}{logical.Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{use.box}{logical.Whether to add box around bootstrap value on tree. Default is TRUE.}

\item{min.ratio}{double (Default:1/30). If min.ratio is not NULL,
all edge length of a phylogenetic tree should be greater than
min.ratio*the longest edge length.
If not, the edge length will be reset as min.ratio*longest edge length.}

\item{show.class.label}{logical. If TRUE, show labels for mutations classes on the heatmap.Default is FALSE.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{plot.geneList}{if TRUE, plot heatmap with genes on geneList when geneList is not NULL.Default FALSE.}

\item{show.gene}{show the name of genes next to the heatmap.Default FALSE.}

\item{show.geneList}{show the names of gene on the geneList.Default FALSE.}

\item{mut.threshold}{show.gene and show.geneList will be FALSE when patient have more mutations than threshold.Default is 150.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{
return a list of phylotree graph .
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree)
## Use CCF 
plotPhyloTree(phyloTree, heatmap.type = 'CCF')
}
