% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree = NULL,
  patient.id = NULL,
  branchCol = "mutSig",
  show.bootstrap = TRUE,
  use.box = TRUE,
  common.col = "red",
  min.ratio = 1/20,
  min.mut.count = 15,
  signaturesRef = "cosmic_v2"
)
}
\arguments{
\item{phyloTree}{PhyloTree object generated by getPhyloTree function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{branchCol}{Character."mutSig" for coloring branches by mutation signature; "mutType" for coloring branches by branch tumor type; NULL for coloring automatically.Default is 'mutSig'.}

\item{show.bootstrap}{Logical. Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{use.box}{Logical. Whether to add box around bootstrap value on tree. Default is TRUE.}

\item{common.col}{Color of common branch which is used for function compareTree.

are mapped along the trees as indicated}

\item{min.ratio}{Double (Default:1/20). If min.ratio is not NULL,
all edge length of a phylogenetic tree should be greater than
min.ratio*the longest edge length.
If not, the edge length will be reset as min.ratio*longest edge length.}

\item{min.mut.count}{The threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}
}
\value{
return a list of phylotree graph .
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree)
}
