% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree = NULL,
  show.mutSig = TRUE,
  show.heatmap = TRUE,
  show.bootstrap = TRUE,
  use.box = TRUE
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getphyloTree function.}

\item{show.mutSig}{logical. Whether to show mutational signature on tree. Default is TRUE.}

\item{show.heatmap}{logical. Whether to show heatmap of somatic mutations. Default is TRUE.}

\item{show.bootstrap}{logical.Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{use.box}{logical.Whether to add box around bootstrap value on tree. Default is TRUE.}
}
\value{
Prints plot.
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree)
## Use ccf 
plotPhyloTree(phyloTree, heatmap.type = 'CCF')
}
