% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{A  phylogenetic tree painter}
\usage{
plotPhyloTree(njtree = NULL, phylotree.type = "njtree",
  use.indel = FALSE, show.mutSig = TRUE, sig.min.mut.number = 50,
  sig.name = "default", show.heatmap = TRUE, heatmap.type = "binary",
  savePlot = FALSE, phylotree.dat = NULL)
}
\arguments{
\item{njtree}{NJtree object}

\item{phylotree.type}{Phylotree format,you can choose "njtree","newick","beast","PAML" with root}

\item{show.mutSig}{if show Mutational Signature in Images}

\item{sig.min.mut.number}{minimum mutation number in each branch}

\item{show.heatmap}{if plot heatmap that show mutation distribution in each branch}

\item{heatmap.type}{type of heatmap,choose 'CCF' if use ccf}

\item{phylotree.dat}{If the format of the phylotree is not "njtree", upload the path of the file to be analyzed with this parameter}
}
\value{
Images of Phylotree
}
\description{
A  phylogenetic tree painter
}
\examples{
plotPhyloTree(njtree)
plotPhyloTree(njtree)
# if use ccf 
plotPhyloTree(njtree, heatmap.type = 'CCF')
# use other tree format
newick.file <- system.file("extdata/newick", "1.nwk", package="MesKit")
plotPhyloTree(phylotree.dat = newick.file1, phylotree.type = 'newick')
beast.file <- system.file("extdata/BEAST", "sample.beast", package="MesKit")
plotPhyloTree(phylotree.dat = beast.file , phylotree.type = 'beast')
PAML.file <- system.file("extdata/PAML", "sample.paml", package="MesKit")
plotPhyloTree(phylotree.dat = PAML.file , phylotree.type = 'PAML')
}
