% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{A  phylogenetic tree painter}
\usage{
plotPhyloTree(njtree = NULL, show.mutSig = TRUE,
  sig.name = "default", show.heatmap = TRUE, heatmap.type = "binary")
}
\arguments{
\item{njtree}{NJtree object}

\item{show.mutSig}{If show Mutational Signature on tree.Default is "True"}

\item{sig.name}{Choose "alias"  when you want showing alias in branch.Default is "default"}

\item{show.heatmap}{If plot heatmap showing mutation distribution in each branch.Default is True}

\item{heatmap.type}{Type of heatmap,choose 'CCF' if use ccf.Default is True}
}
\value{
Images of Phylotree
}
\description{
A  phylogenetic tree painter
}
\examples{
plotPhyloTree(njtree)
plotPhyloTree(njtree)
## Use ccf 
plotPhyloTree(njtree, heatmap.type = 'CCF')
}
