% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloTree.R
\name{plotPhyloTree}
\alias{plotPhyloTree}
\title{plotPhyloTree}
\usage{
plotPhyloTree(
  phyloTree,
  patient.id = NULL,
  branchCol = "mutType",
  show.bootstrap = TRUE,
  min.ratio = 1/20,
  signaturesRef = "cosmic_v2",
  min.mut.count = 15
)
}
\arguments{
\item{phyloTree}{phyloTree or phyloTreeList object generated by getPhyloTree function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{branchCol}{Specify the colors of branches (Default: mutType). Other options: "mutSig" for coloring branches by branch mutation signature;}

\item{show.bootstrap}{Logical. Whether to add bootstrap value on internal nodes.Default is TRUE.}

\item{min.ratio}{Double (Default:1/20). If min.ratio is not NULL,
all edge length of a phylogenetic tree should be greater than
min.ratio*the longest edge length.
If not, the edge length will be reset as min.ratio*longest edge length.}

\item{signaturesRef}{Signature reference,Users can upload their own reference. Default "cosmic_v2". Option:"exome_cosmic_v3","nature2013".}

\item{min.mut.count}{The threshold for the variants in a branch. Default 15.

are mapped along the trees as indicated}
}
\value{
return a list of phylotree graph .
}
\description{
plotPhyloTree
}
\examples{
plotPhyloTree(phyloTree)
}
