% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrunkBranch.R
\name{plotTrunkBranch}
\alias{plotTrunkBranch}
\title{plotTrunkBranch}
\usage{
plotTrunkBranch(tree.mutSig, conf.level = 0.95)
}
\arguments{
\item{tree.mutSig}{the output of function treeMutationalSig.}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95.}
}
\value{
a list of box plots based on mutational categories
}
\description{
Prints the distribution of branch/trunk mutations based on mutational categories.
}
\examples{
plotTrunkBranch(tree.mutSig, conf.level = 0.95)
}
