% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrunkBranch.R
\name{plotTrunkBranch}
\alias{plotTrunkBranch}
\title{plotTrunkBranch}
\usage{
plotTrunkBranch(treeMSOutput, conf.level = 0.95)
}
\arguments{
\item{treeMSOutput}{the output of function treeMutSig.}

\item{conf.level}{confidence level of the interval for wilcox.test. Default: 0.95. Option: on the scale of 0 to 1.}
}
\value{
Box plots based on mutational categories
}
\description{
Prints the distribution of branch/trunk mutations based on mutational categories.
}
\examples{
plotTrunkBranch(treeMSOutput, conf.level = 0.95)
}
