% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{readMaf}
\usage{
readMaf(
  mafFile,
  ccfFile = NULL,
  adjusted.VAF = FALSE,
  mutNonSilent = NULL,
  ccf.conf.level = 0.95,
  refBuild = "hg19"
)
}
\arguments{
\item{mafFile}{Tab delimited MAF file (plain text or *.gz compressed). Required.}

\item{ccfFile}{CCF file of somatic mutations. Default NULL.}

\item{mutNonSilent}{Variant classifications which are considered as non-silent. Default NULL.}

\item{ccf.conf.level}{The confidence level of CCF to identify clonal or subclonal. Only works when "CCF_std" or "CCF_CI_high" is provided in ccfFile. Default: 0.95}

\item{refBuild}{Human reference genome versions of "hg18", "hg19" or "hg38" by UCSC. Default: "hg19".}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSilent".}
}
\value{
an object of class Maf.
}
\description{
Read tab delimited MAF (can be plain text or *.gz compressed) file along with sample information file.
}
\examples{
maf.File <- system.file("extdata/", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC6046.CCF.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
maf <- readMaf(mafFile=maf.File, ccfFile=ccf.File, refBuild="hg19")
}
