% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{readMaf}
\usage{
readMaf(mafFile, sampleInfoFile, mutType = "All", mutNonSilent = NULL,
  chrSilent = NULL, use.indel = FALSE, ccfClusterTsvFile = NULL,
  ccfLociTsvFile = NULL, refBuild = "hg19")
}
\arguments{
\item{mafFile}{tab delimited MAF file (plain text or *.gz compressed). The prefix to the filename of mafFile is considered as "patientID", which will be used for downstream analysis and visualization. Required.}

\item{sampleInfoFile}{sample information file.}

\item{mutType}{Select proper variant classification you need. Default "All".Option: "nonSilent".}

\item{mutNonSilent}{variant classifications which are considered as non-silent. Default NULL. Option: "Default".}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{use.indel}{logic. whether to use INDELs besides somatic SNVs. Default FALSE.}

\item{ccfClusterTsvFile}{CCF file of cluster generated by PyClone. Default NULL.}

\item{ccfLociTsvFile}{CCF file of mutated site generated by PyClone. Default NULL.}

\item{refBuild}{human reference genome versions of hg19 or hg38 by UCSC. Default "hg19". Option: "hg38".}
}
\value{
an object of class Maf.
}
\description{
Read tab delimited MAF (can be plain text or *.gz compressed) file along with sample information file.
}
\examples{
maf.File <- system.file("extdata/maf", "HCC6046.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata", "HCC6046.sampleInfo.txt", package = "MesKit")
pyCloneCluster <- system.file("extdata/ccf", "HCC6046.cluster.tsv", package = "MesKit")
pyCloneLoci <- system.file("extdata/ccf", "HCC6046.loci.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, ccfClusterTsvFile=pyCloneCluster, ccfLociTsvFile=pyCloneLoci, refBuild="hg19")
}
