% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{ReadMaf}
\usage{
readMaf(mafFile, sampleInfoFile, ID = NULL, mutType = "All",
  mutNonSilent = NULL, chrSilent = NULL, use.indel = FALSE,
  ccfClusterTsvFile = NULL, ccfLociTsvFile = NULL, refBuild = "hg19")
}
\arguments{
\item{mafFile}{MAF-format data file.}

\item{sampleInfoFile}{Sample information file.}

\item{ID}{Input your expected patientID character as well as correct the name error of shiny fileinput. Default NULL.}

\item{mutType}{Select proper variant classification you need. Default "All".Option: "nonSilent".}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{use.indel}{A logic parameter to determine whether to use other variant type besides SNP. Default FALSE. Option: TRUE.}

\item{ccfClusterTsvFile}{CCF cluster.tsv file if ccf data provided. Default NULL.}

\item{ccfLociTsvFile}{CCF loci.tsv file if ccf data provided. Default NULL.}

\item{refBuild}{Choose human reference genome versions of hg19 or hg38 by UCSC. Default "hg19". Option: "hg38".}

\item{mutNonSilentAnd}{List variant classifications that you do not want them to be silent.  Default NULL. Option: "Default".}
}
\value{
a Maf object/class
}
\description{
Read tab delimited MAF (can be plain text or gz compressed) file along with sample information file.
}
\examples{
maf.File <- system.file("extdata/maf", "HCC6046.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata", "HCC6046.sampleInfo.txt", package = "MesKit")
pyCloneCluster <- system.file("extdata/ccf", "HCC6046.cluster.tsv", package = "MesKit")
pyCloneLoci <- system.file("extdata/ccf", "HCC6046.loci.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, ccfClusterTsvFile=pyCloneCluster, ccfLociTsvFile=pyCloneLoci, refBuild="hg19")

}
