% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{readMaf}
\usage{
readMaf(
  mafFile,
  ccfFile = NULL,
  adjusted.VAF = FALSE,
  nonSyn.vc = NULL,
  ccf.conf.level = 0.95,
  refBuild = "hg19"
)
}
\arguments{
\item{mafFile}{Tab delimited MAF file (plain text or *.gz compressed). Required.}

\item{ccfFile}{CCF file of somatic mutations. Default NULL.}

\item{adjusted.VAF}{Let VAF = VAF_adj.Default FALSE.}

\item{nonSyn.vc}{List of Variant classifications which are considered as non-silent. Default NULL, use Variant Classifications with "Frame_Shift_Del","Frame_Shift_Ins","Splice_Site","Translation_Start_Site","Nonsense_Mutation","Nonstop_Mutation","In_Frame_Del","In_Frame_Ins","Missense_Mutation"}

\item{ccf.conf.level}{The confidence level of CCF to identify clonal or subclonal. Only works when "CCF_std" or "CCF_CI_high" is provided in ccfFile. Default: 0.95}

\item{refBuild}{Human reference genome versions of "hg18", "hg19" or "hg38" by UCSC. Default: "hg19".}
}
\value{
an object of Maf or MafList.
}
\description{
Read tab delimited MAF (can be plain text or *.gz compressed) file along with sample information file.
}
\examples{
maf.File <- system.file("extdata/", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC6046.CCF.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
maf <- readMaf(mafFile=maf.File, ccfFile=ccf.File, refBuild="hg19")
}
