% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{MAF/CCF/Sample_info Reader}
\usage{
readMaf(mafFile, sampleInfoFile, mut.type = "All",
  mut.nonSilent = NULL, chr.silent = NULL, ccfClusterTsvFile = NULL,
  ccfLociTsvFile = NULL, refBuild = "hg19", inputFileName = NULL)
}
\arguments{
\item{mafFile}{MAF file directory.}

\item{sampleInfoFile}{sample_info.txt file directory.}

\item{mut.type}{select proper variant classification you need}

\item{mut.nonSilent}{}

\item{ccfClusterTsvFile}{CCF cluster.tsv file directory if ccf data provided. Default NULL.}

\item{ccfLociTsvFile}{CCF loci.tsv file directory if ccf data provided. Default NULL.}

\item{refBuild}{BSgenome.Hsapiens.UCSC reference. Default "hg19". Full genome sequences for Homo sapiens (Human) as provided by UCSC.}

\item{inputFileName}{Correct error of shiny app MAF file name}
}
\value{
a classMaf object/class includes information of sample_info and 
mut.id and summary figure of it
}
\description{
Add sample information and ccf data to the original maf file 
to get a new maf. The new maf data adds three pieces of information:lesion, 
patient and time.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "Meskit")
## manually usage
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
## if ccf data provided
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, ccfClusterTsvFile=pyCloneCluster, ccfLociTsvFile=pyCloneLoci, refBuild="hg19")

}
