% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaf.R
\name{readMaf}
\alias{readMaf}
\title{readMaf}
\usage{
readMaf(
  mafFile,
  ccfFile = NULL,
  mutType = "All",
  mutNonSilent = NULL,
  chrSilent = NULL,
  use.indel = FALSE,
  refBuild = "hg19"
)
}
\arguments{
\item{mafFile}{tab delimited MAF file (plain text or *.gz compressed). Required.}

\item{ccfFile}{CCF file of SNVs. Default NULL.}

\item{mutType}{select proper variant classification you need. Default "All".Option: "nonSilent".}

\item{mutNonSilent}{variant classifications which are considered as non-silent. Default NULL.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{use.indel}{logic. whether to use INDELs besides somatic SNVs. Default FALSE.}

\item{refBuild}{human reference genome versions of "hg18", "hg19" or "hg38" by UCSC. Default "hg19".}
}
\value{
an object of class Maf.
}
\description{
Read tab delimited MAF (can be plain text or *.gz compressed) file along with sample information file.
}
\examples{
maf.File <- system.file("extdata/maf", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC6046.CCF.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
maf <- readMaf(mafFile=maf.File, ccfFile=ccf.File, refBuild="hg19")
}
