% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhyloTree.R
\name{readPhyloTree}
\alias{readPhyloTree}
\title{Get sampleID of private branchs and shared branches/trunk in NJtree}
\usage{
readPhyloTree(matTree)
}
\arguments{
\item{matTree}{a nj tree object generated from ape}
}
\value{
a dataframe with sampleID of each branch and trunk
}
\description{
Get sampleID of private branchs and shared branches/trunk in NJtree
}
