% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegment.R
\name{readSegment}
\alias{readSegment}
\title{readSegment}
\usage{
readSegment(
  segCN.file = NULL,
  gtf.dat = NULL,
  gisticAmpGenesFile = NULL,
  gisticDelGenesFile = NULL,
  verbose = TRUE,
  cluster = FALSE,
  min.seg.size = 500
)
}
\arguments{
\item{segCN.file}{character.Path to  copy number file. Column names should be Sample, Chromosome, Start_Position, End_Position, CopyNumber}

\item{gtf.dat}{character.Path to GTF data.Default NULL.}

\item{gisticAmpGenesFile}{Amplification Genes file generated by GISTIC. Default NULL.}

\item{gisticDelGenesFile}{Deletion Genes file generated by GISTIC. Default NULL.}

\item{verbose}{Whether to display details in the console.Default True.}

\item{cluster}{Default FALSE.}

\item{min.seg.size}{Mininum length of segs.Default is 500.}
}
\description{
readSegment
}
\examples{

}
