% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegment.R
\name{readSegment}
\alias{readSegment}
\title{readSegment}
\usage{
readSegment(
  segCN.file = NULL,
  ref.dat = NULL,
  gisticAmpGenesFile = NULL,
  gisticDelGenesFile = NULL,
  verbose = TRUE,
  min.seg.size = 500
)
}
\arguments{
\item{segCN.file}{the segment file}

\item{ref.dat}{the reference genome file.Default NULL.}

\item{gisticAmpGenesFile}{Amplification Genes file generated by GISTIC. Default NULL.}

\item{gisticDelGenesFile}{Deletion Genes file generated by GISTIC. Default NULL.}

\item{verbose}{whether to display details in the console.Default True.}

\item{min.seg.size}{the smallest size of segments.Default is 500.}
}
\description{
readSegment
}
\examples{
segCN.file <- system.file("extdata", "HCC6046.seg.txt", package = "MesKit")
# Load ref.dat
load(system.file("extdata", "Homo_sapiens.GRCh37.75.RData", package = "MesKit"))
gisticAmpGenesFile <- system.file("extdata", "LIHC_amp_genes.conf_99.txt", package = "MesKit")
gisticDelGenesFile <- system.file("extdata", "LIHC_del_genes.conf_99.txt", package = "MesKit")
seg <- readSegment(segCN.file = segCN.file, ref.dat = ref.dat, 
 gisticAmpGenesFile = gisticAmpGenesFile, gisticDelGenesFile = gisticDelGenesFile)
}
