% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegment.R
\name{readSegment}
\alias{readSegment}
\title{readSegment}
\usage{
readSegment(
  segCN.file = NULL,
  gisticAmpGenesFile = NULL,
  gisticDelGenesFile = NULL,
  gisticAllLesionsFile = NULL,
  gistic.qval = 0.25,
  verbose = TRUE,
  min.seg.size = 500
)
}
\arguments{
\item{segCN.file}{The segment file}

\item{gisticAmpGenesFile}{Amplification Genes file generated by GISTIC. Default NULL.}

\item{gisticDelGenesFile}{Deletion Genes file generated by GISTIC. Default NULL.}

\item{gisticAllLesionsFile}{Information of all lesions generated by GISTIC. Default NULL.}

\item{gistic.qval}{The threshold of gistic Q value. Default is 0.25}

\item{verbose}{whether to display details in the console.Default True.}

\item{min.seg.size}{The smallest size of segments.Default is 500.}
}
\description{
readSegment
}
\examples{
segCN.file <- system.file("extdata", "HCC6046.seg.txt", package = "MesKit")
gisticAmpGenesFile <- system.file("extdata", "LIHC_amp_genes.conf_99.txt", package = "MesKit")
gisticDelGenesFile <- system.file("extdata", "LIHC_del_genes.conf_99.txt", package = "MesKit")
gisticAllLesionsFile <- system.file("extdata", "LICH_all_lesions.conf_99.txt", package = "MesKit")
seg <- readSegment(segCN.file = segCN.file,
                   gisticAmpGenesFile = gisticAmpGenesFile,
                    gisticDelGenesFile = gisticDelGenesFile, 
                   gisticAllLesionsFile = gisticAllLesionsFile )

}
