% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMaf.R
\name{subsetMaf}
\alias{subsetMaf}
\title{Subset Maf object}
\usage{
subsetMaf(
  maf,
  geneList = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.vaf = 0,
  max.vaf = 1,
  min.average.vaf = 0,
  min.average.adj.vaf = 0,
  min.ccf = 0,
  min.ref.depth = 0,
  min.alt.depth = 0,
  min.total.depth = 0,
  clonalStatus = NULL,
  use.adjVAF = FALSE
)
}
\arguments{
\item{maf}{Maf object generated by \code{\link{readMaf}} function.}

\item{chrSilent}{Chromosomes excluded in the analysis. e.g, 1, 2, X, Y. Default NULL.}

\item{mutType}{Select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.vaf}{The minimum VAF for filtering variants. Default: 0.}

\item{max.vaf}{The maximum VAF for filtering variants. Default: 1.}

\item{min.average.vaf}{The minimum tumor average VAF for filtering variants. Default: 0.}

\item{min.average.adj.vaf}{The minimum tumor average ajust VAF for filtering variants. Default: 0.}

\item{min.ccf}{The minimum CCF for filtering variants. Default: NULL.}

\item{min.ref.depth}{The minimum reference allele depth for filtering variants. Default: 0.}

\item{min.alt.depth}{The minimum alteratation allele depth for filtering variants. Default: 0.}

\item{min.total.depth}{The minimum total allele depth for filtering variants. Default: 0.}

\item{clonalStatus}{Subset by clonal status.Default: NULL.Option: "Clonal","Subclonal".}

\item{use.adjVAF}{Let VAF = VAF_adj, Tumor_Average_VAF = Tumor_Average_VAF_adj.Default: FALSE.}
}
\value{
Maf data.
}
\description{
Subset Maf object
}
\examples{
maf <- readMaf(mafFile=maf.File, refBuild="hg19")
## Maf
maf_data <- subsetMaf(maf)
}
