% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutationalSig}
\alias{treeMutationalSig}
\title{TreeMutationalSig}
\usage{
treeMutationalSig(
  njtree,
  driverGenesFile = NULL,
  mutThreshold = 50,
  signaturesRef = "cosmic"
)
}
\arguments{
\item{njtree}{a njtree object generated by NJtree function.}

\item{driverGenesFile}{The file with driver gene list. Default NULL.}

\item{mutThreshold}{the threshold for the variants in a branch. Default 50.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic". Option: "nature2013".}
}
\value{
The data frame, treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in njtree object and output the treeMSOutput for further summary and plot functions.
}
\examples{
treeMSOutput <- treeMutationalSig(njtree, driverGenesFile=NULL, mutThreshold=50, signaturesRef="cosmic")
}
