% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutationalSig}
\alias{treeMutationalSig}
\title{Check Mutational Signature for each branch of phylogenetic tree}
\usage{
treeMutationalSig(njtree, driverGenesFile = NULL, mutThreshold = 50,
  signaturesRef = "signatures.cosmic", plot.Signatures = FALSE)
}
\arguments{
\item{njtree}{the njtree object generated by NJtree function.}

\item{driverGenesFile}{the directory of the driver gene list. Default NULL.}

\item{mutThreshold}{the threshold for the variants in a branch. Default 50.}

\item{refBuild}{BSgenome.Hsapiens.UCSC reference. Default "hg19". Full genome sequences for Homo sapiens (Human) as provided by UCSC.}
}
\value{
data frame of each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches
by .treeMutationalBranches and combine the intersection information into njtree
object. Check the mutational signatures for each branch of phylogenetic tree
by using treeMutationalSig and return a data.frame about the most likely muational
signatures.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "Meskit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "Meskit")
maf <- readMaf(patientID = "311252", mafFile = maf.File, sampleInfoFile = sampleInfo.File, refBuild = "hg19")
njtree <- getNJtree(maf)
## general
signature1 <- treeMutationalSig(njtree, mutThreshold=50)
## use driverGenesFile
putativeDriverGenes.File <- system.file("extdata/multi_lesion", "putative_driver_genes.txt", package = "Meskit")
signature2 <- treeMutationalSig(njtree, driverGenesFile=putativeDriverGenes.File, mutThreshold=50)
## use different signature reference
signature3 <- treeMutationalSig(njtree, mutThreshold=50, signaturesRef="signatures.nature2013")

}
