% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutationalSig}
\alias{treeMutationalSig}
\title{TreeMutationalSig}
\usage{
treeMutationalSig(njtree, driverGenesFile = NULL, mutThreshold = 50,
  signaturesRef = "cosmic")
}
\arguments{
\item{njtree}{the njtree object generated by NJtree function.}

\item{driverGenesFile}{the directory of the driver gene list. Default NULL.}

\item{mutThreshold}{the threshold for the variants in a branch. Default 50.}

\item{signaturesRef}{Default "cosmic". Option: "nature2013". the parameter used for deconstructSig.}

\item{plot.signatures}{the parameter used to print the signautre summary plot by ggplot2}

\item{plot.signatures}{choose whether plot signatures}

\item{plot.branchTrunk}{choose whether output a branch-Trunk plot.}

\item{conf.level}{the significance level to vertify whether the kind of mutation is significant.}
}
\value{
data frame of each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches
by .treeMutationalBranches and combine the intersection information into njtree
object. Check the mutational signatures for each branch of phylogenetic tree
by using treeMutationalSig and return a data.frame about the most likely muational
signatures.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "MesKit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "MesKit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "MesKit")
TODO: need to be completed 

}
