% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationalSig.R
\name{treeMutationalSig}
\alias{treeMutationalSig}
\title{TreeMutationalSig}
\usage{
treeMutationalSig(njtree, driverGenesFile = NULL, mutThreshold = 50,
  signaturesRef = "cosmic")
}
\arguments{
\item{njtree}{The njtree object generated by NJtree function.}

\item{driverGenesFile}{The file with driver gene list. Default NULL.}

\item{mutThreshold}{the threshold for the variants in a branch. Default 50.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic". Option: "nature2013".}
}
\value{
The data frame, treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in njtree object and output the treeMSOutput for further summary and plot functions.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "MesKit")
pyCloneCluster <- system.file("extdata/multi_lesion/ccf", "311252.cluster.tsv", package = "MesKit")
pyCloneLoci <- system.file("extdata/multi_lesion/ccf", "311252.loci.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, ccfClusterTsvFile=pyCloneCluster, ccfLociTsvFile=pyCloneLoci, refBuild="hg19")
njtree <- getNJtree(maf)
treeMSOutput <- treeMutationalSig(njtree, driverGenesFile=NULL, mutThreshold=50, signaturesRef="cosmic")

}
