% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMutSig.R
\name{treeMutSig}
\alias{treeMutSig}
\title{treeMutSig}
\usage{
treeMutSig(
  phyloTree,
  min.mut.count = 15,
  signaturesRef = "cosmic_v2",
  associated = NULL,
  signature.cutoff = 0.1,
  withinTumor = FALSE,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree}{PhyloTree object generated by getPhyloTree function.}

\item{min.mut.count}{The threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{Signature reference,Users can upload their own reference. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{associated}{Associated Vector of associated signatures. If given, will narrow the signatures reference to only the ones listed.Default NULL.}

\item{signature.cutoff}{Discard any signature contributions with a weight less than this amount.Default: 0.1.}

\item{withinTumor}{Exploring signatures within tumor. Default: FALSE.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}
}
\value{
A list of data frames, each one contains treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in phyloTree_list or phyloTree object and output the treeMSOutput for further summary and plot functions.
}
\examples{
treeMutSig(phyloTree)
}
