% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMutSig.R
\name{treeMutSig}
\alias{treeMutSig}
\title{treeMutSig}
\usage{
treeMutSig(
  phyloTree_list,
  geneList = NULL,
  min.mut.count = 15,
  signaturesRef = "cosmic_v2",
  associated = c(),
  signature.cutoff = 0.1,
  withinTumor = FALSE,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree_list}{a list of phyloTree object generated by getPhyloTree function.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{min.mut.count}{the threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{Signature reference,Users can upload their own reference. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{associated}{associated Vector of associated signatures. If given, will narrow the signatures reference to only the ones listed.Default NULL.}

\item{signature.cutoff}{Discard any signature contributions with a weight less than this amount.Default: 0.1.}

\item{withinTumor}{Exploring signatures within tumor.Default is FALSE.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{
a list of data frames, each one contains treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in phyloTree_list object and output the treeMSOutput for further summary and plot functions.
}
\examples{
treeMutSig(phyloTree_list, geneList=NULL, min.mut.count=15, signaturesRef="cosmic")
treeMutSig(phyloTree_list, plot = T)
}
