% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMutSig.R
\name{treeMutSig}
\alias{treeMutSig}
\title{treeMutSig}
\usage{
treeMutSig(
  phyloTree,
  geneList = NULL,
  min.mut.count = 15,
  signaturesRef = "cosmic_v2",
  withinType = FALSE,
  plot = TRUE,
  patient.id = NULL
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by phyloTree function.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{min.mut.count}{the threshold for the variants in a branch. Default 15.}

\item{signaturesRef}{The parameter used for deconstructSig. Default "cosmic_v2". Option: "genome_cosmic_v3","exome_cosmic_v3","nature2013".}

\item{plot}{output a list of diagrams that visualize mutational signature of trunk/branches in a phylogenetic tree.Default FALSE.}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{
a list of data frames, each one contains treeMSOutput, containing information about each set/branch's mutational signature.
}
\description{
Figure out the intersection of the variants on different branches 
in phyloTree object and output the treeMSOutput for further summary and plot functions.
}
\examples{
treeMutSig(phyloTree, geneList=NULL, min.mut.count=15, signaturesRef="cosmic")
treeMutSig(phyloTree, plot = T)
}
