% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePathway.R
\name{treePathway}
\alias{treePathway}
\title{treePathway}
\usage{
treePathway(
  phyloTree,
  geneList = NULL,
  pathway.type = "KEGG",
  pval = 0.05,
  pAdjustMethod = "BH",
  qval = 0.2,
  plotType = "dot",
  showCategory = 5,
  patient.id = NULL,
  withinType = FALSE
)
}
\arguments{
\item{phyloTree}{a phyloTree object generated by getPhyloTree function.}

\item{geneList}{list of genes to restrict the analysis. Default NULL.}

\item{pathway.type}{one of "KEGG" or "Reactome". Default type="KEGG"}

\item{pval}{cutoff value of pvalue. Default pval=0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". 
Default is "BH"}

\item{qval}{cutoff value of qvalue. Default qval=0.2}

\item{plotType}{one of "dot", "bar", default is "dot"}

\item{showCategory}{number of categories will be shown, default is 5}

\item{patient.id}{select the specific patients. Default: NULL, all patients are included}
}
\value{
pathway enrichment results
}
\description{
Perform pathway enrichment analysis of trunk/branch mutations of a phylogenetic tree.
}
\examples{
treePathway(phyloTree, pathway.type = "KEGG")
}
