% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMatrix.R
\name{triMatrix}
\alias{triMatrix}
\title{triMatrix}
\usage{
triMatrix(phyloTree, patient.id = NULL, withinTumor = FALSE)
}
\arguments{
\item{phyloTree}{phyloTree or phyloTreeList object generated by getPhyloTree function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included}

\item{withinTumor}{Exploring signatures within tumor. Default: FALSE.}
}
\value{
The frequency of 96 trinucleotide mutation.
}
\description{
Calculate the frequency of 96 trinucleotide mutation based on somatic SNVs.
}
\examples{
triMatrix(phyloTree)
}
