% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMatrix.R
\name{triMatrix}
\alias{triMatrix}
\title{triMatrix}
\usage{
triMatrix(phyloTree, patient.id = NULL, withinTumor = FALSE)
}
\arguments{
\item{phyloTree}{phyloTree or phyloTreeList object generated by \code{\link{getPhyloTree}} function.}

\item{patient.id}{Select the specific patients. Default NULL, all patients are included}

\item{withinTumor}{Exploring signatures within tumor. (Default: FALSE).}
}
\value{
The frequency of 96 trinucleotide mutation.
}
\description{
Calculate the frequency of 96 trinucleotide mutation based on somatic SNVs.
}
\examples{
maf.File <- system.file("extdata/", "HCC_LDC.maf", package = "MesKit")
clin.File <- system.file("extdata/", "HCC_LDC.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "HCC_LDC.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")

## Load a reference genome.
library(BSgenome.Hsapiens.UCSC.hg19)

phyloTree <- getPhyloTree(maf, patient.id = 'HCC8257')
triMatrix(phyloTree)
}
