% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  patient.id = NULL,
  segFile = NULL,
  min.vaf = 0,
  use.adjVAF = FALSE,
  withinTumor = FALSE,
  use.tumorSampleLabel = FALSE,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default NULL, all patients are included.}

\item{segFile}{The segment file.}

\item{min.vaf}{The minimum value of VAF. Default 0.02. Option: on the scale of 0 to 1.}

\item{use.adjVAF}{Use adjusted VAF in analysis when adjusted VAF or CCF is available. Default FALSE.}

\item{withinTumor}{Cluster VAF within tumors in each patients. (Default: FALSE).}

\item{use.tumorSampleLabel}{Logical (Default: FALSE). 
Rename the 'Tumor_Sample_Barcode' by 'Tumor_Sample_Label'.}

\item{...}{Other options passed to \code{\link{subMaf}}}
}
\value{
clustering plots of vaf
}
\description{
Generate variant allele frequency (VAF) frequency distribution curve.
}
\examples{
maf.File <- system.file("extdata/", "CRC_HZ.maf", package = "MesKit")
clin.File <- system.file("extdata/", "CRC_HZ.clin.txt", package = "MesKit")
ccf.File <- system.file("extdata/", "CRC_HZ.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, clinicalFile = clin.File, ccfFile=ccf.File, refBuild="hg19")
vafCluster(maf, patient.id = 'V402')

}
