% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  seg = NULL,
  min.vaf = 0.02,
  max.vaf = 1,
  showMATH = TRUE,
  plotOption = "combine"
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function}

\item{seg}{a segment object generated by readSegment function.}

\item{min.vaf}{the minimum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{max.vaf}{the maximum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{showMATH}{logical. Whether to show MATH Scores in the plot. Default TRUE}

\item{plotOption}{character. Three options for displaying clustering of VAFs in samples. Default "compare"
"compare": prints comparison of VAFs distribution among samples;
"combine" prints VAFs distribution of samples all together;}
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine", showMATH=FALSE)


}
