% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{VAFCluster}
\usage{
vafCluster(maf, minVaf = 0.02, maxVaf = 1, showMATH = TRUE,
  plotOption = "combine", themeOption = "aaas")
}
\arguments{
\item{maf}{The classMaf object generated by readMaf.}

\item{minVaf}{Default: 0. Option: on the scale of 0 to 1. The minimum value of vaf}

\item{maxVaf}{Default: 0. Option: on the scale of 0 to 1. The maximum value of vaf}

\item{showMATH}{Default TRUE. Option: FALSE. Show MATH Scores or hide them.}

\item{plotOption}{choose a vaf plot ouput option for your target data. Default "compare". 
"compare" will tend to provide a ridgeline plot.
"combine" will output a picture with all separate VAF frequency distribution pictures.
"separate" will provide a list of VAF plots forevery samples from the patient.}

\item{themeOption}{Default "aaas". Option: other palette from ggsci. Select a theme palette from ggsci.}
}
\value{
Images of selected samples' VAF
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "MesKit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "MesKit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine")
vafCluster(maf, plotOption="separate")

}
