% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{A VAF plot painter}
\usage{
vafCluster(maf, vafRange = c(0, 1), vafColumn = "VAF",
  plotOption = "compare", themeOption = "aaas", fileFormat = NULL,
  outputDir = NULL)
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{vafRange}{set the VAF threshold for MATH Score calculation. 
The first number in the list given would be the minvaf parameter for mathScore 
and the second one would be the maxvaf parameter. Default c(0, 1).}

\item{plotOption}{choose a vaf plot ouput option for your target data.Default "compare".}

\item{themeOption}{select a theme Palette from ggsci. Default "aaas".}

\item{fileFormat}{decide the ouput format generated by ggsave. Default "pdf".}

\item{outputDir}{set the ouput directory of all outputs. Default NULL.}
}
\value{
Images of selected samples' VAF
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve with ggplot2 as well as ggridges. Different parameters could lead to 
 different outputs with different range of samples.
}
\examples{
## data information
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
maf <- readMaf(patientID="311252", mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
## save ggridge plot
vafCluster(maf, plotOption="311252-S", themeOption="npg", fileFormat="pdf", math_vafRange=c(0,1))

}
