% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  segCN.file = NULL,
  patient.id = NULL,
  chrSilent = NULL,
  mutType = "All",
  use.indel = TRUE,
  min.vaf = 0.02,
  max.vaf = 1,
  showMATH = TRUE,
  plotOption = "combine",
  use.adjVAF = FALSE
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{segCN.file}{the segment file.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{chrSilent}{Select chromosomes needed to be dismissed. Default NULL.}

\item{mutType}{select Proper variant classification you need. Default "All". Option: "nonSyn".}

\item{use.indel}{Logical value. Whether to use INDELs besides somatic SNVs. Default: TRUE.}

\item{min.vaf}{the minimum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{max.vaf}{the maximum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{showMATH}{logical. Whether to show MATH Scores in the plot. Default TRUE}

\item{plotOption}{character. Three options for displaying clustering of VAFs in samples. Default "compare"
"compare": prints comparison of VAFs distribution among samples;
"combine" prints VAFs distribution of samples all together;}

\item{use.adjVAF}{cluster by adjusted VAF.Default: FALSE.}
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine", showMATH=FALSE)


}
