% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  segCN.file = NULL,
  patient.id = NULL,
  min.vaf = 0.02,
  max.vaf = 1,
  plotOption = "combine",
  withinTumor = FALSE,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by readMaf function.}

\item{segCN.file}{The segment file.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{min.vaf}{The minimum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{max.vaf}{The maximum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{plotOption}{Character value. Three options for displaying clustering of VAFs in samples. Default "compare"
"compare": prints comparison of VAFs distribution among samples;
"combine" prints VAFs distribution of samples all together;}

\item{withinTumor}{Cluster VAF within tumors in each patients,default is FALSE.}

\item{...}{Other options passed to \code{\link{subsetMaf}}}
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine")


}
