% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{A VAF plot painter}
\usage{
vafCluster(maf, vafRange = c(0.05, 1), vafColumn = "VAF",
  plotOption = "combine", themeOption = "aaas")
}
\arguments{
\item{maf}{the classMaf object generated by readMaf.}

\item{vafRange}{set the VAF threshold for MATH Score calculation. 
The first number in the list given would be the minvaf parameter for mathScore 
and the second one would be the maxvaf parameter. Default c(0, 1).}

\item{vafColumn}{choose the specific column containing VAF values. Default "VAF".}

\item{plotOption}{choose a vaf plot ouput option for your target data.Default "compare". 
"compare" will tend to provide a ridgeline plot.
"combine" will output a document with all samples of a patient.
"separate" will provide documents for every samples from the patient.}

\item{themeOption}{select a theme Palette from ggsci. Default "aaas".}
}
\value{
Images of selected samples' VAF
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve with ggplot2 as well as ggridges. Different parameters could lead to 
 different outputs with different range of samples.
}
\examples{
maf.File <- system.file("extdata/multi_lesion/maf", "311252.maf", package = "Meskit")
sampleInfo.File <- system.file("extdata/multi_lesion", "sample_info.txt", package = "Meskit")
maf <- readMaf(mafFile=maf.File, sampleInfoFile=sampleInfo.File, refBuild="hg19")
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine")
vafCluster(maf, plotOption="separate")

}
