% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  seg = NULL,
  minVaf = 0.02,
  maxVaf = 1,
  showMATH = TRUE,
  plotOption = "combine",
  themeOption = "aaas"
)
}
\arguments{
\item{maf}{a Maf object generated by readMaf function.}

\item{seg}{object generated by readSegment function.}

\item{minVaf}{the minimum value of VAF. Default: 0. Option: on the scale of 0 to 1.}

\item{maxVaf}{the maximum value of VAF. Default: 0. Option: on the scale of 0 to 1.}

\item{showMATH}{logical. Whether to show MATH Scores in the plot. Default TRUE.}

\item{plotOption}{character. Three options for displaying clustering of VAFs in samples. Default "compare". 
"compare" option prints comparison of VAFs distribution among samples via ggridges.
"combine" option prints multiple plots of VAFs distribution of samples on a page.
"separate" option returns a list of VAFs distribution plots for every sample.}

\item{themeOption}{Select a theme palette from ggsci. Default "aaas". Option: other palette from ggsci.}
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
vafCluster(maf, plotOption="compare")
vafCluster(maf, plotOption="combine", showMATH=FALSE)
vafCluster(maf, plotOption="separate", themeOption="npg")


}
