% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafCluster.R
\name{vafCluster}
\alias{vafCluster}
\title{vafCluster}
\usage{
vafCluster(
  maf,
  patient.id = NULL,
  segFile = NULL,
  min.vaf = 0.02,
  max.vaf = 1,
  withinTumor = FALSE,
  ...
)
}
\arguments{
\item{maf}{Maf or MafList object generated by \code{\link{readMaf}} function.}

\item{patient.id}{Select the specific patients. Default: NULL, all patients are included.}

\item{segFile}{The segment file.}

\item{min.vaf}{The minimum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{max.vaf}{The maximum value of VAF. Default: 0. Option: on the scale of 0 to 1}

\item{withinTumor}{Cluster VAF within tumors in each patients,default is FALSE.}

\item{...}{Other options passed to \code{\link{subMaf}}}
}
\value{
clustering plots of vaf
}
\description{
Generate variant allele frequency(VAF) frequency distribution
 curve in different mode.
}
\examples{
maf.File <- system.file("extdata", "HCC6046.maf", package = "MesKit")
ccf.File <- system.file("extdata", "HCC6046.ccf.tsv", package = "MesKit")
maf <- readMaf(mafFile=maf.File, ccfFile = ccf.File, refBuild="hg19")
vafCluster(maf)

}
